### 2025-01-26 #7721 @warrickball
Hi @GiMattia, thanks for your submission. All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, note that your paper is currently too long. The paper [shouldn't be much more than 1000 words say](https://joss.readthedocs.io/en/latest/paper.html). I encourage you to have a look at [what a JOSS paper should contain](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) and remove a lot of the detail that should be part of the project documentation. If you want to regenerate the paper here, you can comment with `@editorialbot generate pdf` if you aren't already running our preprint pipeline yourself.

Also note that [we encourage hyperlinks](https://joss.readthedocs.io/en/latest/paper.html#links) over bare URLs, including in (and sometimes instead of) footnotes.
