+++ Paper/paper.bib
@@ -170,13 +170,19 @@ archivePrefix = {arXiv},
     url          = {https://doi.org/10.5281/zenodo.3509134}
 }
 
-@InCollection{ParaView,
-  author    = {James Ahrens and Berk Geveci and Charles Law},
-  booktitle = {Visualization Handbook},
-  publisher = {Elesvier},
-  title     = {{ParaView}: An End-User Tool for Large Data Visualization},
-  year      = {2005},
-  note      = {{ISBN}~978-0123875822},
+@inbook{ParaView,
+title = "ParaView: An end-user tool for large-data visualization",
+abstract = "This chapter describes the design and features of a visualization tool called ParaView, a tool that allows scientists to visualize and analyze extremely large datasets. The tool provides a graphical user interface for the creation and dynamic execution of visualization tasks. ParaView transparently supports the visualization and rendering of large datasets by executing these programs in parallel on shared or distributed memory machines. ParaView supports hardware-accelerated parallel rendering and achieves interactive rendering performance via level-of-detail techniques. The design balances and integrates a number of diverse requirements, including the ability to handle large data, ease of use, and extensibility by developers. The chapter describes the requirements that guided the design, identifies the importance of those requirements to scientific users, and discusses key design decisions and tradeoffs. {\textcopyright} 2005",
+author = "Ahrens, {James Paul} and Berk Geveci and Charles Law",
+year = "2005",
+month = dec,
+day = "1",
+doi = "10.1016/B978-012387582-2/50038-1",
+language = "English",
+isbn = "9780123875822",
+pages = "717--731",
+booktitle = "Visualization Handbook",
+publisher = "Unknown Publisher",
 }
 
 @ARTICLE{Puzzoni2021,
