### 2025-02-03 b781a73
fixed typo path figures
