### 2025-02-03 #7721 @GiMattia
Hi @warrickball, Thank you very much for the prompt reply. We shortened the paper by almost entirely removing the "main features" section. If it is still too long, we may remove the examples as well (although I will first check if now the paper is short enough without removing them if it is ok).
We have also replaced the URLs with hyperlinks, being able to get rid of almost all the footnotes.
