### 2025-02-03 #7721 @warrickball
Hi @GiMattia, this is an excellent start, thanks! Though we'd ideally reduce the word count a bit further, I happy for the review to proceed as is and I can make some suggestions towards the end of the review process. The content of the paper might change during the review anyway.
