### 2025-03-19 5ce0942
added grid in params set_axis
