+++ Src/pyPLUTO/__init__.py
@@ -6,7 +6,7 @@ from .loadpart import LoadPart
 from .pytools import ring, savefig, show
 
 # Define the version and additional environment variables
-__version__ = "1.0"
+__version__ = "1.0.0"
 __colorerr__ = True
 __colorwarn__ = True
 __session__ = find_session()
