### 2025-04-24 #7721 @GiMattia
Hi, I was wondering if there is any update or, if not, something like an estimate on the time needed to get an editor.
