+++ .github/workflows/test_linux.yml
@@ -12,7 +12,6 @@ jobs:
     strategy:
       matrix:
         os:
-          - ubuntu-20.04
           - ubuntu-22.04
           - ubuntu-24.04
         python-version: ["3.10", "3.11", "3.12", "3.13"]
