### 2025-04-28 48a97c3
started refactoring and cleaning code
