+++ Docs/source/conf.py
@@ -7,7 +7,6 @@ import os
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here.
-import pathlib
 import sys
 
 sys.path.insert(0, os.path.abspath("../../Src/pyPLUTO/"))
@@ -15,10 +14,14 @@ sys.path.insert(0, os.path.abspath("../../Src/pyPLUTO/"))
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "PyPLUTO"
-copyright = "2024, G. Mattia, D. Crocco, D. Melon Fuksman, M. Bugli, V. Berta, E. Puzzoni, A. Mignone, B, Vaidya"
-author = "G. Mattia D. Crocco, D. Melon Fuksman, M. Bugli, V. Berta, E. Puzzoni, A. Mignone, B, Vaidya"
-release = "1.0"
+project: str = "PyPLUTO"
+copyright: str = (
+    "2024, G. Mattia, D. Crocco, D. Melon Fuksman, M. Bugli, V. Berta, E. Puzzoni, A. Mignone, B, Vaidya"
+)
+author: str = (
+    "G. Mattia D. Crocco, D. Melon Fuksman, M. Bugli, V. Berta, E. Puzzoni, A. Mignone, B, Vaidya"
+)
+release: str = "1.0"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
