+++ Src/pyPLUTO/load.py
@@ -270,7 +270,7 @@ class Load:
         # Find the format of the data files
         self._find_format(datatype, (alone := kwargs.get("alone", None)))
 
-        # Find relevant informations without opening the files (e.g.
+        # Find relevant information without opening the files (e.g.
         # the number of files to be loaded) or opening the *.out files
         if self._alone is True:
             self._findfiles(nout)
@@ -303,7 +303,8 @@ class Load:
             print(f"Load: folder {path},     output {_nout_out}")
 
         # Try to read the file definitions.h
-        if defh := kwargs.get("defh", None) is not False:
+        defh = kwargs.get("defh", None)
+        if defh is not False:
             pathdefh = self.pathdir / "definitions.h"
             defhfile = "definitions.hpp"
             if not pathdefh.exists():
@@ -315,13 +316,13 @@ class Load:
                 print(f"No {defhfile} is read!") if defh is True else ...
 
         # Try to read the file pluto.ini
-        if plini := kwargs.get("plini", None) is not False:
+        plini = kwargs.get("plini", None)
+        if plini is not False:
             pathplini = self.pathdir / "pluto.ini"
             try:
                 setattr(self, "plini", self._read_plini(pathplini))
             except:
                 print(f"No pluto.ini is read!") if plini is True else ...
-
         return
 
     def __str__(self):
