### 2025-04-28 #7721 @warrickball
Hi @GiMattia, sorry again for the long wait. The Astronomy, Astrophysics & Space Science (AASS) track of JOSS has a substantial backlog, which is typically leading to a ~3 month wait from submission to an editor becoming available. PyPLUTO is now very close to the front of the queue so hopefully an editor with appropriate expertise will pick this up soon.

I've personally been away for most of April and many editors and reviewers will have had breaks over Easter. I can't offer any guarantee but I'd hope an editor can pick this up and start looking for reviewers in the next few weeks.
