### 2025-04-29 8c8a844
started refactoring and cleaning code, tests + GUI (ruff)
