+++ Examples/test09_blast.py
@@ -10,92 +10,61 @@ $PLUTO_DIR/Test_Problems/MHD/Blast (configuration 9).
 The data is loaded twice into a pload object D and the Image class is created.
 The slices method is used to obtain a slice of the desired quantity along the
 diagonals of the domain. The plot method and the legend method are then used
-to plot an highly informative plot of density and pression with customized
+to plot a highly informative plot of density and pression with customized
 legend labels. The image is then saved and shown on screen.
 """
 
 # Loading the relevant packages
 import numpy as np
+import pyPLUTO
 
-import pyPLUTO as pp
+# Initialization
+Image = pyPLUTO.Image(suptitle="Test 09 - MHD Blast test")
+Image.create_axes(ncol=2)
 
-# Creating the Image, the subplot axes and defining the colors of the lines
-I = pp.Image(suptitle="Test 09 - MHD Blast test")
-I.create_axes(ncol=2)
-col = I.color[:2]
 
-# Loading the initial data into a pload object D
-D = pp.Load(0, path="Test_Problems/MHD/Blast")
+# Helper function to plot a frame
+def plot_frame(Data, ax_idx: int, time_label: str):
+    x = Data.x1 * np.sqrt(2)
 
-# Plotting the data in the first subplot
-varM = D.slices(D.rho, diag=True)
-varm = D.slices(D.rho, diag="min")
+    for idx, (var, label, color, yrange) in enumerate(
+        [
+            (Data.rho, r"$\rho$", Image.color[0], [0.1, 200]),
+            (Data.prs, r"$p$", Image.color[1], None),
+        ]
+    ):
+        var_max = Data.slices(var, diag=True)
+        var_min = Data.slices(var, diag="min")
+        Image.plot(
+            x,
+            var_max,
+            c=color,
+            ax=ax_idx,
+            label=label,
+            yscale="log" if label == r"$p$" else "linear",
+        )
+        Image.plot(
+            x,
+            var_min,
+            c=color,
+            ax=ax_idx,
+            ls="--",
+            yrange=yrange,
+            title=f"t = {time_label} s",
+            xtitle="x",
+        )
 
-I.plot(
-    D.x1 * np.sqrt(2),
-    varM,
-    c=col[0],
-    ax=0,
-    label=r"$\rho$",
-    legpos=2,
-    legsize=13,
-)
+    Image.legend(ax=ax_idx, legpos=1, label=["M", "m"], ls=["-", "--"])
 
-I.plot(D.x1 * np.sqrt(2), varm, c=col[0], ax=0, ls="--", yrange=[0.1, 200])
 
-varM = D.slices(D.prs, diag=True)
-varm = D.slices(D.prs, diag="min")
+# Plotting the initial time
+Data = pyPLUTO.Load(0, path="Test_Problems/MHD/Blast")
+plot_frame(Data, ax_idx=0, time_label="0.0")
 
-I.plot(D.x1 * np.sqrt(2), varM, c=col[1], ax=0, yscale="log", label=r"$p$")
+# Plotting the final time
+Data = pyPLUTO.Load(path="Test_Problems/MHD/Blast")
+plot_frame(Data, ax_idx=1, time_label="0.01")
 
-I.plot(
-    D.x1 * np.sqrt(2),
-    varm,
-    c=col[1],
-    ax=0,
-    ls="--",
-    xtitle=r"$x$",
-    title="t = 0.0 s",
-)
-
-I.legend(ax=0, legpos=1, label=["M", "m"], ls=["-", "--"])
-
[…]
