### 2025-04-29 f4a4ce7
started refactoring and cleaning code, amr (ruff)
