+++ Src/pyPLUTO/3dplot.py
@@ -1,6 +1,2 @@
-from .libraries import *
-
 # WIP
 raise NotImplementedError("WIP: 3D plotting has not been implemented yet.")
-
-# comment
