### 2025-04-29 ba0c525
started refactoring and cleaning code, defh plini echo (ruff)
