### 2025-04-29 cffcc74
started refactoring and cleaning code, figure / find_fieldlines (ruff)
