### 2025-04-29 d11450b
started refactoring and cleaning code, h_image/pypluto (ruff)
