+++ Src/pyPLUTO/h_image.py
@@ -1,4 +1,8 @@
-from .libraries import *
+from matplotlib.axes import Axes
+from numpy.typing import NDArray
+import numpy as np
+import warnings
+from typing import Any
 
 
 def _add_ax(self, ax: Axes, i: int) -> None:
