### 2025-04-29 2de3683
started refactoring and cleaning code, image interact (ruff)
