### 2025-04-29 5b5e5d5
started refactoring and cleaning code, load (fluid/part) (ruff)
