### 2025-04-29 8fd6f8a
started refactoring and cleaning code, parttools / nabla (ruff)
