### 2025-04-29 15dca4e
started refactoring and cleaning code, plots/zoom/lines (ruff)
