+++ Src/pyPLUTO/plotzoom.py
@@ -1,4 +1,8 @@
-from .libraries import *
+from matplotlib.axes import Axes
+from .h_pypluto import check_par
+from typing import Any
+import numpy as np
+import warnings
 
 
 def zoom(self, ax=None, check: bool = True, **kwargs) -> Axes:
