### 2025-04-29 b113f6a
started refactoring and cleaning code, tools (ruff)
