### 2025-04-29 9c52d7c
ruff load class
