+++ Src/pyPLUTO/readdata.py
@@ -1,4 +1,7 @@
-from .libraries import *
+from .h_pypluto import makelist
+import numpy as np
+from pathlib import Path
+import warnings
 
 
 def _load_variables(
