+++ requirements.txt
@@ -1,10 +1,0 @@
-pyqt6~=6.9.0
-pyqt6-sip~=13.10.0
-matplotlib~=3.10.1
-numpy~=2.2.5
-contourpy~=1.3.2
-pandas~=2.2.3
-h5py~=3.13.0
-scipy~=1.15.2
-setuptools~=75.8.0
-pytest~=8.3.5

[…]
