+++ Src/pyPLUTO/image.py
@@ -172,7 +172,7 @@ class Image:
 
     def __getattr__(self, name):
         try:
-            return getattr(self, f"_{name}")
+            return object.__getattribute__(self, f"_{name}")
         except AttributeError:
             raise AttributeError(f"'Image' object has no attribute '{name}'")
 
