+++ Src/pyPLUTO/3dplot.py
@@ -1,3 +1,2 @@
 # WIP
 raise NotImplementedError("WIP: 3D plotting has not been implemented yet.")
-
