+++ Tests/test_im01_initimage.py
@@ -39,7 +39,6 @@ def test_fontsize():
     assert Image.fontsize == 20
 
 
-
 # Test the tight layout
 def test_tight_layout():
     Image = pp.Image(tight=False)
