+++ Src/pyPLUTO/3dplot.py
@@ -1,2 +1,0 @@
-# WIP
-raise NotImplementedError("WIP: 3D plotting has not been implemented yet.")
