### 2025-04-29 0417d25
removed trash, added packages to project.toml, started enhancing coverage
