### 2025-04-30 75cbf03
1st refactoring step: image_new!
