+++ Src/pyPLUTO/__init__.py
@@ -1,5 +1,6 @@
 # Import the libraries, classes and functions
 from .image import Image
+from .image_new import Image_new
 import warnings
 import sys
 from .h_pypluto import find_session, color_warning, color_error
@@ -12,7 +13,7 @@ __version__ = "1.0"
 __colorerr__ = True
 __colorwarn__ = True
 __session__ = find_session()
-__all__ = ["Image", "Load", "LoadPart", "ring", "savefig", "show"]
+__all__ = ["Image", "Image_new", "Load", "LoadPart", "ring", "savefig", "show"]
 
 # Print the version and session
 print(f"PyPLUTO version: {__version__}   session: {__session__}")
