+++ Examples/test09_blast.py
@@ -16,6 +16,7 @@ legend labels. The image is then saved and shown on screen.
 
 # Loading the relevant packages
 import numpy as np
+
 import pyPLUTO
 
 # Initialization
