+++ Tests/test_lf04_loadout.py
@@ -1,5 +1,6 @@
 import os
 from pathlib import Path
+
 import numpy as np
 import numpy.testing as npt
 
