+++ Src/pyPLUTO/load.py
@@ -1,8 +1,10 @@
-from typing import Any
-from numpy.typing import NDArray
-from .h_pypluto import check_par
 from pathlib import Path
+from typing import Any
+
 import numpy as np
+from numpy.typing import NDArray
+
+from .h_pypluto import check_par
 
 
 class Load:
