+++ pyproject.toml
@@ -51,3 +51,11 @@ dev = [
 
 [tool.setuptools.packages.find]
 where = ["Src"]
+
+[tool.isort]
+profile = "black"
+line_length = 88
+multi_line_output = 3
+include_trailing_comma = true
+force_grid_wrap = 0
+use_parentheses = true
