+++ .github/workflows/pre-commit.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Install pre-commit and dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pre-commit black flake8 isort
+          pip install pre-commit black flake8 isort ruff
 
       # Run pre-commit hooks
       - name: Run pre-commit
