+++ .github/workflows/pre-commit.yml
@@ -1,34 +1,0 @@
-name: Check code style
-
-on:
-  pull_request:
-  push:
-    branches:
-      - master
-
-jobs:
-  pre-commit:
-    runs-on: ubuntu-latest
-
-    steps:
-      # Checkout the repository
-      - name: Checkout repository
-        uses: actions/checkout@v3
-
-      # Set up Python environment
-      - name: Set up Python
-        uses: actions/setup-python@v3
-        with:
-          python-version: 3.12
-
-      # Install dependencies
-      - name: Install pre-commit and dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install pre-commit black flake8 isort ruff
-
-      # Run pre-commit hooks
-      - name: Run pre-commit
-        uses: pre-commit/action@v3.0.0
-        with:
-          extra_args: --all-files
