+++ .github/workflows/code-style.yml
@@ -1,29 +1,0 @@
-# .github/workflows/code-style.yml
-name: Check code Style
-
-on:
-  pull_request:
-  push:
-    branches: [main]
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: '3.13'
-
-      - name: Install dependencies
-        run: |
-          pip install ruff black isort mypy
-
-      - name: Ruff check
-        run: ruff check .
-
-      - name: Black check
-        run: black --check .
-
-      - name: Isort check
-        run: isort --check-only .
