+++ .github/workflows/pre-commit.yml
@@ -20,6 +20,9 @@ jobs:
         uses: actions/setup-python@v3
         with:
           python-version: 3.12
+          
+      - name: Clear cache
+        run: rm -rf ~/.cache/pip
 
       # Install dependencies
       - name: Install pre-commit and dependencies
