+++ .pre-commit-config.yaml
@@ -22,13 +22,13 @@ repos:
         args: ["--markdown-linebreak-ext=md"]
 
   # 2. isort import
-  - repo: https://github.com/pycqa/isort
-    rev: "6.0.1"
-    hooks:
-      - id: isort
-        args: ["--profile", "black"]
-        files: \.py$
-        stages: [pre-commit]
+  #- repo: https://github.com/pycqa/isort
+ #   rev: "6.0.1"
+  #  hooks:
+ #     - id: isort
+   #     args: ["--profile", "black"]
+   #     files: \.py$
+   #     stages: [pre-commit]
 
   # 3. Format with Black
   - repo: https://github.com/psf/black
