+++ .github/workflows/test_macos.yml
@@ -1,4 +1,4 @@
-name: Run Tests on MacOS
+name: Tests on MacOS
 
 on:
   push:
