+++ .pre-commit-config.yaml
@@ -4,14 +4,22 @@ repos:
     rev: v5.0.0
     hooks:
       - id: check-added-large-files
+        files: \.py$
       - id: check-case-conflict
+        files: \.py$
       - id: check-merge-conflict
+        files: \.py$
       - id: debug-statements
+        files: \.py$
       - id: end-of-file-fixer
+        files: \.py$
       - id: mixed-line-ending
+        files: \.py$
       - id: trailing-whitespace
+        files: \.py$
         args: ["--markdown-linebreak-ext=md"]
 
+        
   # 1. Code spelling
   #- repo: https://github.com/codespell-project/codespell
   #  rev: 'v2.4.1'
