+++ Tests/test_im03_set_axis.py
@@ -1,6 +1,5 @@
 import numpy as np
 import numpy.testing as npt
-
 import pyPLUTO as pp
 
 
