+++ Tests/test_im02_createaxis.py
@@ -1,5 +1,4 @@
 import numpy as np
-
 import pyPLUTO as pp
 
 
