+++ Src/pyPLUTO/plot1d.py
@@ -15,8 +15,7 @@ def plot(
     check: bool = True,
     **kwargs: Any,
 ) -> None:
-    """
-    Creation of a 1D function plot (or a 1D slice plot).
+    """Creation of a 1D function plot (or a 1D slice plot).
     It creates a simple figure and a single axis if none are given prior.
     If a single function argument is given, it plots the graph of that function
     using a linear variable as x parameter. However, if a pair of arrays is
@@ -24,12 +23,10 @@ def plot(
 
     Returns
     -------
-
     - None
 
     Parameters
     ----------
-
     - alpha: float, default 1.0
         Sets the opacity of the plot, where 1.0 means total opaque and 0.0 means
         total transparent.
@@ -166,14 +163,12 @@ def plot(
 
     Notes
     -----
-
     - Minorticks on single axis will be added in future releases.
 
     ----
 
     Examples
-    ========
-
+    --------
     - Example #1: create a simple plot of y as function of x
 
         >>> import pyPLUTO as pp
@@ -213,7 +208,6 @@ def plot(
         >>> I.plot(x, z, ax = I.ax[0])
 
     """
-
     # If only one argument is given, it is the y-axis
     if y[0] is None:
         y = np.asarray(x)
@@ -304,7 +298,7 @@ def plot(
     # Creation of the legend
     self.legpos[nax] = kwargs.get("legpos", self.legpos[nax])
     if self.legpos[nax] is not None:
-        copy_label: str | None = kwargs.get("label", None)
+        copy_label: str | None = kwargs.get("label")
         kwargs["label"] = None
         legend(self, ax, check=False, fromplot=True, **kwargs)
         kwargs["label"] = copy_label
@@ -314,7 +308,6 @@ def plot(
         self.fig.tight_layout()
 
     # End of the function
-    return None
 
 
 def legend(
@@ -324,15 +317,13 @@ def legend(
     fromplot: bool = False,
     **kwargs: Any,
 ) -> None:
-    """
-    Creation of a legend referring to the current figure.
+    """Creation of a legend referring to the current figure.
 
     If no labels are given, it shows the labels of all the plots in the figure,
     ordered by entry. If specific labels are given, it shows those ones.
 
     Returns
     -------
-
     - None
 
     Parameters
@@ -384,14 +375,12 @@ def legend(
 
     Notes
     -----
-
     - None
 
     ----
 
     Examples
-    ========
-
+    --------
     - Example #1: create a standard legend
 
         >>> import pyPLUTO as pp
@@ -421,7 +410,6 @@ def legend(
         ... label = ['continue', 'dotted'])
 
     """
-
     # Check parameters
     param = {
         "c",
@@ -518,4 +506,3 @@ def legend(
     ax.add_artist(legg)
 
     # End of the function
-    return None
