+++ Examples/test12_riemannlp.py
@@ -20,7 +20,9 @@ import pyPLUTO
 
 # Loading the data and the particle data into two pload objects
 Data_f = pyPLUTO.Load(path="Test_Problems/Particles/LP/Riemann_2D")
-Data_p = pyPLUTO.LoadPart(path="Test_Problems/Particles/LP/Riemann_2D", nfile_lp=0)
+Data_p = pyPLUTO.LoadPart(
+    path="Test_Problems/Particles/LP/Riemann_2D", nfile_lp=0
+)
 
 # Creating the image and the axes
 Image = pyPLUTO.Image(figsize=[9.5, 8])
@@ -28,7 +30,9 @@ Image.create_axes(right=0.8)
 Image.create_axes(nrow=2, left=0.81, right=0.85, hspace=0.01)
 
 # Plotting the data
-pcm = Image.display(Data_f.rho, x1=Data_f.x1, x2=Data_f.x2, ax=0, aspect="equal")
+pcm = Image.display(
+    Data_f.rho, x1=Data_f.x1, x2=Data_f.x2, ax=0, aspect="equal"
+)
 indx = np.argsort(Data_p.vx1**2 + Data_p.vx2**2 + Data_p.vx3**2)
 
 # Plot the fluid colorbar
