+++ Src/pyPLUTO/plotzoom.py
@@ -352,7 +352,9 @@ def _zoomdisplay(self, ax: Axes, nax: int, axins: Axes, **kwargs: Any) -> None:
     kwargs["vmax"] = kwargs.pop("vmax", self.vlims[nax][1])
     kwargs["tresh"] = kwargs.pop("tresh", self.vlims[nax][2])
 
-    self.display(pcm0, x1=xc, x2=yc, ax=axins, check=False, shading=psh, **kwargs)
+    self.display(
+        pcm0, x1=xc, x2=yc, ax=axins, check=False, shading=psh, **kwargs
+    )
 
 
 def _place_inset_pos(ax: Axes, pos: list[float]) -> Axes:
