+++ GUI/config.py
@@ -79,7 +79,9 @@ def select_folder(self):
         datatype = datatype if datatype in formats_list else None
         datatry = file_path.split("/")[-1].split(".")[0]
         self.datatype = (
-            datatry if datatry in formats_list and datatype is None else datatype
+            datatry
+            if datatry in formats_list and datatype is None
+            else datatype
         )
         try:
             self.nout = int(file_path.split("/")[-1].split(".")[1])
