### 2025-05-08 a7e8870
added tests created delegator decorator
