### 2025-05-08 cea4766
removed unnecessay text from delgator
