### 2025-05-08 5795290
added mypy.txt file to do strict check
