+++ mypy_files.txt
@@ -1,0 +1,4 @@
+Src/pyPLUTO/image_new.py
+Src/pyPLUTO/figure_new.py
+Src/pyPLUTO/delegator.py
+Src/pyPLUTO/imagestate.py
