### 2025-05-08 47fafc3
added test_figure.py
