### 2025-05-10 4533ad0
added automatic inspect_kwargs
