+++ Src/pyPLUTO/image_new.py
@@ -6,11 +6,13 @@ from matplotlib.figure import Figure
 from .delegator import delegator
 from .figure_new import FigureManager
 from .imagestate import ImageState
+from .inspect_kwargs import track_kwargs
 
 
 @delegator("state")
 class Image_new:
 
+    @track_kwargs
     def __init__(
         self,
         LaTeX: bool | str = True,
@@ -21,5 +23,10 @@ class Image_new:
         **kwargs: Any,
     ) -> None:
 
-        self.state: ImageState = ImageState(style=style, fig=fig or Figure())
-        self.figure_manager: FigureManager = FigureManager(self.state)
+        self.state: ImageState = ImageState(
+            style=style, LaTeX=LaTeX, fig=fig or Figure()
+        )
+        self.figure_manager: FigureManager = FigureManager(self.state, **kwargs)
+
+        if text:
+            print("Image class created at ")  # NEED TO FINISH THE SENTENCE
