+++ mypy_files.txt
@@ -2,3 +2,4 @@ Src/pyPLUTO/image_new.py
 Src/pyPLUTO/figure_new.py
 Src/pyPLUTO/delegator.py
 Src/pyPLUTO/imagestate.py
+Src/pyPLUTO/inspect_kwargs.py
