+++ .github/workflows/coverage.yml
@@ -21,7 +21,7 @@ jobs:
         run: |
             python -m pip install --upgrade pip
             pip install pytest pytest-cov
-            pip install -r requirements.txt
+            pip install -r requirements_dev.txt
       - name: Test and Coverage
         run: python -m pytest --cov --cov-report json
 
