+++ .github/workflows/coverage.yml
@@ -10,18 +10,26 @@ on:
 jobs:
   test:
     runs-on: ubuntu-latest
+
     steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
       # Your original steps
       - name: Set up Python 3.13
-        uses: actions/setup-python@v3
+        uses: actions/setup-python@v4
         with:
             python-version: "3.13"
 
-      - name: Install
+      - name: Install dependencies
         run: |
-            python -m pip install --upgrade pip
-            pip install pytest pytest-cov
-            pip install -r requirements_dev.txt
+          python -m pip install --upgrade pip
+          pip install -r requirements_dev.txt
+
+      - name: Install pyPLUTO
+        run: |
+          python -m pip install .
+
       - name: Test and Coverage
         run: python -m pytest --cov --cov-report json
 
