### 2025-05-11 14b4dfe
fixed tests path + added coverage
