+++ Tests/test_lf02_loadformat.py
@@ -6,6 +6,7 @@ import pytest
 
 # Assuming the root of the repo is your current working directory
 repo_root = Path(os.getcwd())
+repo_root = repo_root if repo_root.name == "Tests" else repo_root / "Tests"
 path = repo_root / "Test_load"
 
 
@@ -76,7 +77,7 @@ def test_multiple_findformat():
 # Given format (multiple files), alone = True
 def test_multiple_alone():
     D = pp.Load(
-        path="Test_load/multiple_files", text=False, datatype="vtk", alone=True
+        path=path / "multiple_files", text=False, datatype="vtk", alone=True
     )
     assert D.format == "vtk"
 
