+++ coverage.txt
@@ -1,74 +1,0 @@
-============================= test session starts ==============================
-platform linux -- Python 3.12.7, pytest-8.3.4, pluggy-1.5.0
-rootdir: /home/gian/WIP/pyPLUTO
-configfile: pyproject.toml
-plugins: doctestplus-1.3.0, cov-6.0.0, anyio-4.6.2.post1, approvaltests-14.5.0
-collected 128 items
-
-Tests/test_figure_new.py ....                                            [  3%]
-Tests/test_im01_createfig.py ...........                                 [ 11%]
-Tests/test_im02_createaxis.py .........                                  [ 18%]
-Tests/test_im03_set_axis.py .......                                      [ 24%]
-Tests/test_im04_plot.py ..........                                       [ 32%]
-Tests/test_im05_legend.py ...                                            [ 34%]
-Tests/test_im06_display.py .                                             [ 35%]
-Tests/test_im08_zoomplot.py .....                                        [ 39%]
-Tests/test_image_new.py .......                                          [ 44%]
-Tests/test_imagestate.py ....                                            [ 47%]
-Tests/test_lf01_loadpath.py .....                                        [ 51%]
-Tests/test_lf02_loadformat.py ............                               [ 60%]
-Tests/test_lf03_loadgrid.py ...                                          [ 63%]
-Tests/test_lf04_loadout.py ....                                          [ 66%]
-Tests/test_lf05_findfiles.py ...                                         [ 68%]
-Tests/test_lf06_varsoffset.py .........                                  [ 75%]
-Tests/test_lf07_readdata.py ..............                               [ 86%]
-Tests/test_lp01_loadpath.py ....                                         [ 89%]
-Tests/test_lp02_loadformat.py ......                                     [ 94%]
-Tests/test_lp03_findfiles.py .                                           [ 95%]
-Tests/test_lp04_varsoffset.py ...                                        [ 97%]
-Tests/test_lp05_readdata.py ...                                          [100%]
-
----------- coverage: platform linux, python 3.12.7-final-0 -----------
-Name                             Stmts   Miss  Cover
-----------------------------------------------------
-Src/pyPLUTO/__init__.py             19      0   100%
-Src/pyPLUTO/amr.py                 276    271     2%
-Src/pyPLUTO/defpluto.py             35     24    31%
-Src/pyPLUTO/delegator.py            38      2    95%
-Src/pyPLUTO/echo_load.py            37     35     5%
-Src/pyPLUTO/figure.py               57     22    61%
-Src/pyPLUTO/figure_new.py           23      2    91%
-Src/pyPLUTO/findlines.py           154    142     8%
-Src/pyPLUTO/fourier.py              34     29    15%
-Src/pyPLUTO/h_image.py              89     16    82%
-Src/pyPLUTO/h_pypluto.py            34     15    56%
-Src/pyPLUTO/image.py                60      0   100%
-Src/pyPLUTO/image_new.py            14      1    93%
-Src/pyPLUTO/imagestate.py           14      0   100%
-Src/pyPLUTO/imagetools.py           20     14    30%
-Src/pyPLUTO/inspect_kwargs.py       48      8    83%
-Src/pyPLUTO/interact.py             66     58    12%
-Src/pyPLUTO/load.py                154     19    88%
-Src/pyPLUTO/loadpart.py             71      6    92%
-Src/pyPLUTO/make_docstrings.py      40     40     0%
-Src/pyPLUTO/nabla.py               247    237     4%
-Src/pyPLUTO/parttools.py            31     25    19%
-Src/pyPLUTO/plot1d.py               63      0   100%
[…]
