+++ .github/workflows/coverage.yml
@@ -4,8 +4,6 @@ name: pytest-coverage-comment
 
 on:
     push:
-      branches:
-        - master
     pull_request:
 
 # https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs
