### 2025-05-11 5fd1dd4
added other badges 1st try
