+++ .pre-commit-config.yaml
@@ -30,15 +30,7 @@ repos:
   #      files: \.py$
   #      stages: [pre-commit]
 
-  # 2. Black - code formatter
-  - repo: https://github.com/psf/black
-    rev: 25.1.0
-    hooks:
-      - id: black
-        files: \.py$
-        stages: [pre-commit]
-
-  # 3. Blackened-docs (formats code in docstrings)
+  # 2. Blackened-docs (formats code in docstrings)
   - repo: https://github.com/adamchainz/blacken-docs
     rev: 1.19.1
     hooks:
@@ -48,7 +40,7 @@ repos:
         additional_dependencies:
         - black==22.1.0
 
-  # 4. Docformatter (formats docstrings)
+  # 3. Docformatter (formats docstrings)
   - repo: https://github.com/PyCQA/docformatter
     rev: v1.7.7
     hooks:
@@ -57,12 +49,19 @@ repos:
         files: \.py$
         stages: [pre-commit]
 
-  # 5. Ruff - linter, fixer (but NOT formatter)
+  # 4. Ruff - linter, fixer (but NOT formatter)
   - repo: https://github.com/astral-sh/ruff-pre-commit
     rev: v0.11.7
     hooks:
       - id: ruff
         args: ["--fix"]
+        files: \.py$
+        stages: [pre-commit]
+  # 5. Black - code formatter
+  - repo: https://github.com/psf/black
+    rev: 25.1.0
+    hooks:
+      - id: black
         files: \.py$
         stages: [pre-commit]
 
