+++ .github/workflows/coverage_tests.yml
@@ -9,7 +9,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [latest]
+        python-version: ["3.13"]
 
     steps:
       - name: Checkout code
