+++ .github/workflows/coverage_tests.yml
@@ -23,7 +23,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements_devs.txt
+          pip install -r requirements_dev.txt
 
       - name: Install pyPLUTO
         run: |
