+++ pyproject.toml
@@ -112,6 +112,27 @@ omit = [
     "*/site-packages/*"
 ]
 
+[tool.pylint.main]
+# Paths to analyze (you can still override in the CLI if needed)
+jobs = 1
+persistent = true
+suggestion-mode = true
+exit-zero = true
+
+[tool.pylint.messages_control]
+disable = [
+    "missing-docstring",
+    "invalid-name",
+    "too-few-public-methods",
+    "too-many-arguments",
+]
+
+[tool.pylint.format]
+max-line-length = 81
+
+[tool.pylint.reports]
+output-format = "text"
+
 [tool.interrogate]
 verbose = 1
 fail-under = 80
