+++ .github/workflows/coverage_tests.yml
@@ -29,5 +29,14 @@ jobs:
         run: |
           python -m pip install .
 
+      - name: Run Tests with Coverage
+        run: |
+            coverage erase
+            coverage run -m pytest
+      - name: Report Coverage
+        run: |
+            coverage report --show-missing
+            coverage xml -o Docs/Badges/coverage.xml
+
       - name: Run static checks (coverage, pylint, interrogate)
         run: python coverage_tests.py
