+++ pyproject.toml
@@ -133,6 +133,7 @@ max-line-length = 81
 [tool.pylint.reports]
 output-format = "text"
 
+
 [tool.interrogate]
 verbose = 1
 fail-under = 80
