+++ Docs/Badges/pylint.txt
@@ -1,0 +1,848 @@
+************* Module pyPLUTO.plot2d
+Src/pyPLUTO/plot2d.py:734:0: C0301: Line too long (83/81) (line-too-long)
+Src/pyPLUTO/plot2d.py:13:0: R0914: Too many local variables (19/15) (too-many-locals)
+Src/pyPLUTO/plot2d.py:254:14: W0212: Access to a protected member _assign_ax of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:271:4: W0212: Access to a protected member _hide_text of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:285:11: W0212: Access to a protected member _set_cscale of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:314:0: R0914: Too many local variables (17/15) (too-many-locals)
+Src/pyPLUTO/plot2d.py:509:14: W0212: Access to a protected member _assign_ax of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:519:4: W0212: Access to a protected member _hide_text of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:542:11: W0212: Access to a protected member _set_cscale of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:677:13: W0212: Access to a protected member _assign_ax of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:690:20: W0212: Access to a protected member _assign_ax of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:691:8: W0212: Access to a protected member _hide_text of a client class (protected-access)
+Src/pyPLUTO/plot2d.py:716:0: R0917: Too many positional arguments (6/5) (too-many-positional-arguments)
+Src/pyPLUTO/plot2d.py:778:15: E1123: Unexpected keyword argument 'linthresh' in constructor call (unexpected-keyword-arg)
+Src/pyPLUTO/plot2d.py:779:9: R1714: Consider merging these comparisons with 'in' by using 'cscale in ('twoslope', '2slope')'. Use a set instead if elements are hashable. (consider-using-in)
+Src/pyPLUTO/plot2d.py:784:15: E1123: Unexpected keyword argument 'linear_width' in constructor call (unexpected-keyword-arg)
+Src/pyPLUTO/plot2d.py:717:4: W0613: Unused argument 'self' (unused-argument)
+************* Module pyPLUTO.make_docstrings
+Src/pyPLUTO/make_docstrings.py:27:16: W0621: Redefining name 'keys' from outer scope (line 60) (redefined-outer-name)
+Src/pyPLUTO/make_docstrings.py:30:15: W0718: Catching too general exception Exception (broad-exception-caught)
+Src/pyPLUTO/make_docstrings.py:34:12: W0621: Redefining name 'name' from outer scope (line 60) (redefined-outer-name)
+Src/pyPLUTO/make_docstrings.py:51:19: W0718: Catching too general exception Exception (broad-exception-caught)
+Src/pyPLUTO/make_docstrings.py:45:24: W0612: Unused variable 'ispkg' (unused-variable)
+************* Module pyPLUTO.plotlines
+Src/pyPLUTO/plotlines.py:11:0: R0914: Too many local variables (23/15) (too-many-locals)
+Src/pyPLUTO/plotlines.py:197:14: W0212: Access to a protected member _assign_ax of a client class (protected-access)
+Src/pyPLUTO/plotlines.py:210:4: W0212: Access to a protected member _hide_text of a client class (protected-access)
[…]
