+++ .github/workflows/coverage_tests.yml
@@ -29,14 +29,9 @@ jobs:
         run: |
           python -m pip install .
 
-      - name: Run Tests with Coverage
-        run: |
-            coverage erase
-            coverage run -m pytest
-      - name: Report Coverage
-        run: |
-            coverage report --show-missing
-            coverage xml -o Docs/Badges/coverage.xml
+      - name: Run Tests
+        run: pytest -v -cov=Src/pyPLUTO -cov-report=xml -cov-report=term-missing
+        working-directory: ./Tests
 
       - name: Run static checks (coverage, pylint, interrogate)
         run: python coverage_tests.py
