+++ .github/workflows/coverage_tests.yml
@@ -37,7 +37,7 @@ jobs:
      #   run: python coverage_tests.py
 
       - name: Upload badge files as artifacts
-        uses: actions/upload-artifact@v3
+        uses: actions/upload-artifact@v4
         with:
           name: badges
           path: Docs/Badges/*.svg
