+++ .github/workflows/coverage_tests.yml
@@ -1,43 +1,41 @@
-name: Badge Generator
-
+# This workflow will install dependencies, create coverage tests and run Pytest Coverage Comment
+# For more information see: https://github.com/MishaKav/pytest-coverage-comment/
+name: pytest-coverage-comment
 on:
   push:
   pull_request:
 
+# https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs
+# `contents` is for permission to the contents of the repository.
+# `pull-requests` is for permission to pull request
+permissions:
+  contents: write
+  checks: write
+  pull-requests: write
+
 jobs:
-  generate-badges:
+  build:
     runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: ["3.13"]
-
     steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
+      - uses: actions/checkout@v4
 
-      - name: Set up Python
+      - name: Set up Python 3.13
         uses: actions/setup-python@v4
         with:
-          python-version: ${{ matrix.python-version }}
+          python-version: 3.13
 
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -r requirements_dev.txt
+          pip install flake8 pytest pytest-cov
+          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
 
-      - name: Install pyPLUTO
+      - name: Build coverage file
         run: |
-          python -m pip install .
+          pytest --junitxml=pytest.xml --cov-report=term-missing:skip-covered --cov=app tests/ | tee pytest-coverage.txt
 
-      - name: Run Tests
-        run: pytest -v -cov=Src/pyPLUTO -cov-report=xml -cov-report=term-missing
-        working-directory: ./Tests
-
-      #- name: Run static checks (coverage, pylint, interrogate)
-     #   run: python coverage_tests.py
-
-      - name: Upload badge files as artifacts
-        uses: actions/upload-artifact@v4
+      - name: Pytest coverage comment
+        uses: MishaKav/pytest-coverage-comment@main
         with:
-          name: badges
-          path: Docs/Badges/*.svg
+          pytest-coverage-path: ./pytest-coverage.txt
+          junitxml-path: ./pytest.xml
