+++ .github/workflows/coverage_tests.yml
@@ -37,5 +37,31 @@ jobs:
       - name: Pytest coverage comment
         uses: MishaKav/pytest-coverage-comment@main
         with:
-          pytest-coverage-path: ./pytest-coverage.txt
-          junitxml-path: ./pytest.xml
+          pytest-xml-coverage-path: ./coverage.xml
+      
+      - name: Check the output coverage
+        run: |
+            echo "Coverage Percentage - ${{ steps.coverageComment.outputs.coverage }}"
+            echo "Coverage Color - ${{ steps.coverageComment.outputs.color }}"
+            echo "Coverage Html - ${{ steps.coverageComment.outputs.coverageHtml }}"
+  
+            echo "Coverage Warnings - ${{ steps.coverageComment.outputs.warnings }}"
+  
+            echo "Coverage Errors - ${{ steps.coverageComment.outputs.errors }}"
+            echo "Coverage Failures - ${{ steps.coverageComment.outputs.failures }}"
+            echo "Coverage Skipped - ${{ steps.coverageComment.outputs.skipped }}"
+            echo "Coverage Tests - ${{ steps.coverageComment.outputs.tests }}"
+            echo "Coverage Time - ${{ steps.coverageComment.outputs.time }}"
+  
+            echo "Not Success Test Info - ${{ steps.coverageComment.outputs.notSuccessTestInfo }}"
+  
+      - name: Create the Badge
+        uses: schneegans/dynamic-badges-action@v1.7.0
+        with:
+          auth: ${{ secrets.PYTEST_COVERAGE_COMMENT }}
+          gistID: 5e90d640f8c212ab7bbac38f72323f80
+          filename: pytest-coverage-comment__main.json
+          label: Coverage Report
+          message: ${{ steps.coverageComment.outputs.coverage }}
+          color: ${{ steps.coverageComment.outputs.color }}
+          namedLogo: python
