+++ .github/workflows/coverage_tests.yml
@@ -32,7 +32,7 @@ jobs:
 
       - name: Build coverage file
         run: |
-          pytest --junitxml=pytest.xml --cov=pyPLUTO --cov-report=term-missing:skip-covered Tests/
+          pytest --junitxml=pytest.xml --cov=Src/pyPLUTO --cov-report=pytest-coverage.txt  --cov-report=term-missing:skip-covered Tests/
 
       - name: Pytest coverage comment
         uses: MishaKav/pytest-coverage-comment@main
