+++ .github/workflows/coverage_tests.yml
@@ -32,7 +32,8 @@ jobs:
 
       - name: Build coverage file
         run: |
-          pytest --junitxml=pytest.xml --cov=Src/pyPLUTO --cov-report=pytest-coverage.txt  --cov-report=term-missing:skip-covered Tests/
+          pytest --junitxml=pytest.xml --cov=Src/pyPLUTO --cov-report=term-missing:skip-covered Tests/  | tee pytest-coverage.txt
+
 
       - name: Pytest coverage comment
         uses: MishaKav/pytest-coverage-comment@main
