+++ .github/workflows/coverage_tests.yml
@@ -61,7 +61,7 @@ jobs:
         uses: schneegans/dynamic-badges-action@v1.7.0
         with:
           auth: ${{ secrets.PYTEST_COVERAGE_COMMENT }}
-          gistID: 5e90d640f8c212ab7bbac38f72323f80
+          gistID: 384b1f3a3a3b74cdbd65c4e3dce0632f
           filename: pytest-coverage-comment__main.json
           label: Coverage Report
           message: ${{ steps.coverageComment.outputs.coverage }}
