+++ .github/workflows/coverage_tests.yml
@@ -46,8 +46,7 @@ jobs:
         run: |
             echo "Coverage Percentage - ${{ steps.coverageComment.outputs.coverage }}"
             echo "Coverage Color - ${{ steps.coverageComment.outputs.color }}"
-            echo "Coverage Html - ${{ steps.coverageComment.outputs.coverageHtml }}"
-  
+      
             echo "Coverage Warnings - ${{ steps.coverageComment.outputs.warnings }}"
   
             echo "Coverage Errors - ${{ steps.coverageComment.outputs.errors }}"
