+++ .github/workflows/test_coverage.yml
@@ -3,8 +3,8 @@
 name: pytest-coverage-comment
 on:
   push:
- #   branches:
-  #    - master
+    branches:
+      - master
   pull_request:
 
 # https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs
