### 2025-05-13 a018851
try pylint logo
