+++ .github/workflows/test_pylint.yml
@@ -2,8 +2,8 @@ name: pylint-score
 
 on:
     push:
-    #  branches:
-    #    - master
+      branches:
+        - master
     pull_request:
 
 permissions:
