+++ .github/workflows/test_interrogate.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Generate interrogate badge
         run: |
             mkdir -p temp-badges
-            interrogate Src/ --fail-under=0 --badge-path Docs/interrogate.svg
+            interrogate Src/ --fail-under=0 --generate-badge Docs/interrogate.svg
   
       - name: Upload badge to Gist
         uses: schneegans/dynamic-badges-action@v1.7.0
