### 2025-05-14 f23de36
refactored figure.py + tests
