+++ .github/workflows/test_coverage.yml
@@ -1,10 +1,11 @@
 # This workflow will install dependencies, create coverage tests and run Pytest Coverage Comment
 # For more information see: https://github.com/MishaKav/pytest-coverage-comment/
 name: pytest-coverage-comment
+
 on:
   push:
-    branches:
-      - master
+  #  branches:
+  #    - master
   pull_request:
 
 # https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs
