### 2025-05-14 6eb82e8
removed one redundant latex test
