### 2025-05-14 d04ecb9
removed one redundant latex test
