### 2025-05-14 9c9e5df
changed docstring length + added create figurd in image new
