+++ Src/pyPLUTO/load.py
@@ -11,9 +11,9 @@ class Load:
     """The Load class loads the data (fluid) from the output files. The
     initialization corresponds to the loading, if wanted, of one or more
     datafiles for the fluid. The data are loaded in a memory mapped numpy
-    multidimensional array. Such approach does not load the full data until
-    needed. Basic operations (i.e. no numpy) are possible, as well as slicing
-    the arrays, without fully loading the data.
+    multidimensional array. Such approach does not load the full data
+    until needed. Basic operations (i.e. no numpy) are possible, as well
+    as slicing the arrays, without fully loading the data.
 
     Returns
     -------
