### 2025-05-18 6e20ce2
Added docstrings and tests to image_new
