+++ Tests/test_figure_new.py
@@ -42,7 +42,7 @@ def test_interaction_with_image_new():
     assert img.state.LaTeX == True
     assert img.state.style == "default"  # style should fallback to default
     assert (
-        img.figure_manager.state.style == "default"
+        img._figure_manager.state.style == "default"
     )  # Same for the figure manager
 
 
@@ -54,15 +54,15 @@ def test_interaction_with_image_new_valid_style():
         img.state.style == "ggplot"
     )  # style should remain as seaborn-darkgrid
     assert (
-        img.figure_manager.state.style == "ggplot"
+        img._figure_manager.state.style == "ggplot"
     )  # Same for the figure manager
 
 
 def test_number_colors():
     # Given
     img = Image_new(numcolors=15)
-    assert img.figure_manager.state.color[0] == "#0104fe"
-    assert img.figure_manager.color[0] == "#0104fe"
+    assert img._figure_manager.state.color[0] == "#0104fe"
+    assert img._figure_manager.color[0] == "#0104fe"
     with pytest.warns(
         DeprecationWarning,
         match="numcolor is deprecated. Use numcolors instead.",
@@ -75,12 +75,12 @@ def test_latex():
     img = Image_new(LaTeX=False)
     assert img.LaTeX == False
     assert img.state.LaTeX == False
-    assert img.figure_manager.state.LaTeX == False
+    assert img._figure_manager.state.LaTeX == False
 
     img = Image_new(LaTeX=True)
     assert img.LaTeX == True
     assert img.state.LaTeX == True
-    assert img.figure_manager.state.LaTeX == True
+    assert img._figure_manager.state.LaTeX == True
 
 
 def test_latex_pgf(monkeypatch):
@@ -95,7 +95,7 @@ def test_latex_pgf(monkeypatch):
     # Now this should pass
     assert img.LaTeX == "pgf"
     assert img.state.LaTeX == "pgf"
-    assert img.figure_manager.state.LaTeX == "pgf"
+    assert img._figure_manager.state.LaTeX == "pgf"
 
 
 def test_latex_pgf_latex_not_installed(monkeypatch):
@@ -107,7 +107,7 @@ def test_latex_pgf_latex_not_installed(monkeypatch):
 
     with pytest.warns(UserWarning, match="LaTeX not installed"):
         img = Image_new(LaTeX="pgf")  # triggers fallback inside constructor
-    #    img.figure_manager._assign_LaTeX("normal")
+    #    img._figure_manager._assign_LaTeX("normal")
 
     assert img.state.LaTeX is True  # fallback occurred
     img = Image_new(LaTeX=True)
@@ -146,7 +146,7 @@ def test_latex_true_font_missing(monkeypatch):
     with pytest.warns(
         UserWarning, match="LaTeX = True option is not available"
     ):
-        img.figure_manager._assign_LaTeX("bold")
+        img._figure_manager._assign_LaTeX("bold")
 
 
 def test_latex_true_success(monkeypatch):
@@ -155,7 +155,7 @@ def test_latex_true_success(monkeypatch):
     # No warning expected if assignment is valid
     with warnings.catch_warnings(record=True) as w:
         warnings.simplefilter("always")
-        img.figure_manager._assign_LaTeX("normal")
+        img._figure_manager._assign_LaTeX("normal")
         assert len(w) == 0
 
 
@@ -165,7 +165,7 @@ def test_figure_created():
 
     # Then
     assert img.fig is not None
-    assert img.figure_manager.fig is not None
+    assert img._figure_manager.fig is not None
     assert isinstance(img.fig, mpl.figure.Figure)
     assert img.fig.get_figwidth() == 8.0
     assert img.fig.get_figheight() == 5.0
