### 2025-05-18 08e5723
deleted some folders
