+++ Src/pypluto.egg-info/SOURCES.txt
@@ -1,40 +1,0 @@
-LICENSE
-README.md
-pyproject.toml
-Src/pyPLUTO/3dplot.py
-Src/pyPLUTO/__init__.py
-Src/pyPLUTO/amr.py
-Src/pyPLUTO/echo_load.py
-Src/pyPLUTO/figure.py
-Src/pyPLUTO/findlines.py
-Src/pyPLUTO/fourier.py
-Src/pyPLUTO/h_image.py
-Src/pyPLUTO/h_pypluto.py
-Src/pyPLUTO/image.py
-Src/pyPLUTO/imagetools.py
-Src/pyPLUTO/interact.py
-Src/pyPLUTO/libraries.py
-Src/pyPLUTO/load.py
-Src/pyPLUTO/loadpart.py
-Src/pyPLUTO/nabla.py
-Src/pyPLUTO/parttools.py
-Src/pyPLUTO/plot1d.py
-Src/pyPLUTO/plot2d.py
-Src/pyPLUTO/plotlines.py
-Src/pyPLUTO/plotzoom.py
-Src/pyPLUTO/pytools.py
-Src/pyPLUTO/read_files.py
-Src/pyPLUTO/readdata.py
-Src/pyPLUTO/readfluid.py
-Src/pyPLUTO/readformat.py
-Src/pyPLUTO/readgridout.py
-Src/pyPLUTO/readpart.py
-Src/pyPLUTO/setaxes.py
-Src/pyPLUTO/transform.py
-Src/pyPLUTO/trash.py
-Src/pyPLUTO/write_files.py
-Src/pypluto.egg-info/PKG-INFO
-Src/pypluto.egg-info/SOURCES.txt
-Src/pypluto.egg-info/dependency_links.txt
-Src/pypluto.egg-info/requires.txt
-Src/pypluto.egg-info/top_level.txt
