### 2025-05-21 6667667
added skeleton setaxes_new
