### 2025-05-23 5306684
added Config class to initilize the module
