### 2025-05-26 0ee9cbc
added docstrings
