### 2025-05-30 #7721 @arfon
:wave: @GiMattia, I'll be your editor for this review. Could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?
