### 2025-06-03 #7721 @GiMattia
Hi, sorry for the late reply. I was on vacation.
Thank you very much for being the editor.
Based on the management system, I think the following people would be good candidates to review this submission due to the Python language, the data visualization topic, and Astrophysics as an area of expertise:
- falconstryker
- tnehezd
- rgmyr
- ngoldbaum
- FaustinCarter

I did not put the "@" to avoid tagging them and disturbing them. I hope it is okay.
If none of them are available, I can remove some of the constraints.
