### 2025-06-16 6a8d78a
added imagetools & range
