+++ Src/pyPLUTO/imagetools_new.py
@@ -1,0 +1,327 @@
+from collections.abc import Sequence
+from typing import Any
+
+from matplotlib.axes import Axes
+from matplotlib.text import Text
+
+from .create_axes import CreateAxesManager
+from .delegator import delegator
+from .imagestate import ImageState
+from .inspector import track_kwargs
+
+
+@delegator("state")
+class ImageToolsManager:
+
+    exposed_methods = (
+        "savefig",
+        "text",
+        "show",
+    )
+
+    def __init__(self, state: ImageState):
+        """Initialize the ImageToolsManager with the given state."""
+        self.state = state
+        self.AxesManager = CreateAxesManager(state)
+
+    def savefig(
+        self,
+        filename: str = "img.png",
+        bbox: str | None = "tight",
+        dpi: int = 300,
+    ) -> None:
+        """Creation of a .png image file of the figure created with the
+        Image class.
+
+        Returns
+        -------
+        - None
+
+        Parameters
+        ----------
+        - bbox: {'tight', None}, default 'tight'
+            Crops the white borders of the Image to create a more balanced image
+            file.
+        - filename: str, default 'img.png'
+            The name of the saved image file.
+
+        Notes
+        -----
+        - None
+
+        ----
+
+        Examples
+        --------
+        - Example #1: save an empty image
+
+            >>> import pyPLUTO as pp
+            >>> I = pp.Image()
+            >>> I.savefig('namefile.png')
+
+        """
+        if not self.state.fig:
+            raise ValueError("No figure to save. Please create a figure first.")
+        self.state.fig.savefig(filename, bbox_inches=bbox, dpi=dpi)
+
+    def show(
+        self,
+    ) -> None:
+        raise NotImplementedError(
+            "Image show is deprecated, please use pp.show instead"
+        )
+
+    @track_kwargs
+    def text(
+        self,
+        text: str,
+        x: float = 0.85,
+        y: float = 0.85,
+        ax: Axes | None = None,
+        check: bool = True,
+        **kwargs: Any,
+    ) -> None:
+        """Insertion of a text box inside the figure created with Image
+        class.
+
+        Returns
+        -------
+        - None
+
+        Parameters
+        ----------
+        - ax: axis object, default None
+            The axis where to insert the text box. If None, the last considered
+            axis will be used.
+        - c: str, default 'k'
+            Determines the text color.
+        - horalign: str, default 'left'
+            The horizontal alignment. Possible values are 'left', 'center',
+            'right'.
+        - text (not optional): str
+            The text that will appear on the text box
+        - textsize: float, default fontsize
+            Sets the text fontsize. The default value corresponds to the value
+            of the actual fontsize in the figure.
+        - veralign: str, default 'baseline'
+            The vertical alignment. Possible values are 'baseline', 'bottom',
+            'center', 'center_baseline', 'top'.
+        - x: float, default 0.85
+            The horizontal starting position of the text box, in units of figure
+            size.
+        - xycoords: str, default 'fraction'
+            The coordinate system used. Possible values are 'figure fraction',
+            which sets the position as a fraction of the axis (inside the axis
+            lie values between 0 and 1), 'points', which sets the position in
+            units of the x/y coordinate system, and 'figure', which sets the
+            position as a fraction of the figure.
+        - y: float, default 0.85
+            The vertical starting position of the text box, in units of figure
+            size.
+
+        Notes
+        -----
+        - None
+
+        ----
+
+        Examples
+        --------
+        - Example #1: Insert text inside a specific axis
+
+            >>> I.text('text', x = 0.5, y = 0.5, ax = ax)
+
[…]
