### 2025-06-19 #7721 @tnehezd
Hi @GiMattia,

To be honest, this will be my first JOSS review, so I'm quite new to the process. Above, I've accepted the review invitation, I've read the guidelines.

@arfon, @warrickball: Just to confirm, do I need to initiate something on my end, or will you be sending me direct access/instructions for this review?

Regards,
Dóri
