### 2025-06-20 #7721 @warrickball
Hi @tnehezd,

> Just to confirm, do I need to initiate something on my end, or will you be sending me direct access/instructions for this review?

Not yet. When the editor starts the review, you'll get the instructions on how to conduct the review. You can have a look at any recent JOSS review to see what this currently looks like. e.g. https://github.com/openjournals/joss-reviews/issues/8419.
