### 2025-06-23 #8448 @editorialbot
[REVIEW]: PyPLUTO: a data analysis Python package for the PLUTO code
**Submitting author:** @GiMattia (<a href="http://orcid.org/0000-0003-1454-6226">Giancarlo Mattia</a>)
**Repository:** https://github.com/GiMattia/PyPLUTO/
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
