### 2025-06-23 #7721 @falconstryker
> @falconstryker @tnehezd @rgmyr @ngoldbaum @FaustinCarter – :wave: would any of you be willing to review this submission for JOSS? The submission under consideration is _[PyPLUTO: a data analysis Python package for the PLUTO code](https://github.com/openjournals/joss-reviews/issues/7721#top)_
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. You can learn more about the process in these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> Based on your experience, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!
>
> Many thanks
> Arfon

Apologies, I've been sick for the last week. Happy to participate in future reviews but will have to pass on this one.
