+++ Src/pyPLUTO/interactive.py
@@ -1,0 +1,409 @@
+from collections.abc import Iterable
+from typing import Any
+
+import matplotlib.pyplot as plt
+import numpy as np
+from matplotlib import animation
+from matplotlib.artist import Artist
+from matplotlib.axes import Axes
+from matplotlib.collections import Collection, QuadMesh
+from matplotlib.lines import Line2D
+from matplotlib.widgets import Slider
+from numpy.typing import NDArray
+
+from .delegator import delegator
+from .display import DisplayManager
+from .imagestate import ImageState
+from .imagetools_new import ImageToolsManager
+from .inspector import track_kwargs
+from .plot_new import PlotManager
+
+
+@delegator("state")
+class InteractiveManager:
+
+    exposed_methods = (
+        "animate",
+        "interactive",
+    )
+
+    def __init__(self, state: ImageState):
+
+        self.state = state
+        self.DisplayManager = DisplayManager(state)
+        self.ImageToolsManager = ImageToolsManager(state)
+
+        self.PlotManager = PlotManager(state)
+        self.anim_pcm: Collection | Line2D | None = None
+        self.labslider: list[str | float] | None = None
+        self.anim_ax: Axes | None = None
+        self.anim_var: dict[str, NDArray[Any]] | NDArray[Any] | None = None
+        self.animkeys: NDArray[Any] | None = None
+        self.nsld: int = 0
+        self.lenlab: int = 0
+        self.limfix: bool = True
+        self.slider: Slider | None = None
+
+    @track_kwargs
+    def interactive(
+        self,
+        varx: dict[str, NDArray[Any]] | NDArray[Any],
+        vary: dict[str, NDArray[Any]] | None = None,
+        check: bool = True,
+        limfix: bool = True,
+        labslider: list[str | float] | None = None,
+        **kwargs: Any,
+    ) -> None:
+        """Creates an interactive plot with a slider to change the data.
+        Warning: it works only with the fluid variables.
+
+        Returns
+        -------
+        - None
+
+        Parameters
+        ----------
+        - varx (not optional): array_like
+            The x-axis variable.
+        - vary: array_like, default None
+            The y-axis variable.
+        - ax: Axes, default None
+            The axes instance.
+        - labslider: str, default None
+            The label of the slider.
+        - limfix: bool, default True
+            If True, the colorbar limits are fixed through the entire animation.
+        - **kwargs: Any
+            Other parameters to pass used in the plot or display functions.
+        - vmin: float, default None
+            The minimum value of the data.
+        - vmax: float, default None
+            The maximum value of the data.
+
+        Notes
+        -----
+        - Check vmin and vmax to set the colorbar limits.
+        - Interact is a very primordial version and it should not be used tu
+            perform very complex plots. Instead, it gives a very nice overview
+            of the data as functions of time (like in other softwares such as
+            visit or paraview).
+
+        ----
+
+        Examples
+        --------
+        - Example #1: Create an interactive 2D plot
+
+            >>> import pyPLUTO as pp
+            >>> D = pp.Load('all')
+            >>> I = pp.Image()
+            >>> I.interactive(D.rho, x1 = D.x1, x2 = D.x2,
+            ... cpos = 'right', vmin = 0, vmax = 1.0)
+            ...
+            >>> pp.show()
+
+        - Example #2: Create an interactive 1D plot with a composite variable
+
+            >>> import pyPLUTO as pp
+            >>> import numpy as np
+            >>> D = pp.Load('all')
+            >>> pp.Image().interactive(D.x1, np.sqrt(D.vx1**2 + D.vx2**2))
+            >>> pp.show()
+
+        """
+        kwargs.pop("check", check)
+
[…]
