### 2025-07-01 f017373
finished image refactoring
