+++ Src/pyPLUTO/plotlines.py
@@ -1,547 +1,0 @@
-import warnings
-from typing import Any
-
-import numpy as np
-from matplotlib.collections import LineCollection
-from numpy.typing import NDArray
-
-from .h_pypluto import check_par
-
-
-def contour(
-    self, var: NDArray, check: bool = True, **kwargs: Any
-) -> LineCollection:
-    """Plots a contour plot of a given variable. The function uses the
-    matplotlib.pyplot.contour function. The function returns None.
-
-    Returns
-    -------
-    - cnt: LineCollection
-        The set of contour lines of the given variable.
-
-    Parameters
-    ----------
-    - alpha: float, default 1.0
-        Sets the transparency of the contour lines.
-    - aspect: {'auto', 'equal', float}, default 'auto'
-        Sets the aspect ratio of the plot. The 'auto' keyword is the default
-        option (most likely the plot will be squared). The 'equal' keyword will
-        set the same scaling for x and y. A float will fix the ratio between the
-        y-scale and the x-scale (1.0 is the same as 'equal').
-    - ax: {ax object, 'old', None}, default None
-        The axis where to plot the lines. If None, a new axis is created.
-        If 'old', the last considered axis will be used.
-    - c: str, default ['k','#12e3c0','#3f6600','#1815c5','#f67451','#d7263d']
-        Determines the contour lines plot. If not defined, the program will loop
-        over an array of 6 color which are different for the most common vision
-        deficiencies.
-    - cmap: str, default 'hot'
-        Selects the colormap. If not defined, the colormap 'hot' will be
-        adopted. Some useful colormaps are: plasma, magma, seismic. Please avoid
-        using colorbars like jjet or rainbow, which are not perceptively uniform
-        and not suited for people with vision deficiencies.
-        All the colormap available are listed in the following link:
-        https://matplotlib.org/stable/tutorials/colors/colormaps.html
-    - cpos: {'top','bottom','left','right'}, default None
-        Enables the colorbar (if defined), default position on the right.
-    - cscale: {'linear','log','symlog','twoslope'}, default 'linear'
-        Sets the colorbar scale. Default is the linear ('norm') scale.
-    - extend: {'neither','both','min','max'}, default 'neither'
-        Sets the extension of the triangular colorbar extension.
-    - extendrect: bool, default False
-        If True, the colorbar extension will be rectangular.
-    - fontsize: float, default 17.0
-        Sets the fontsize for all the axis components (only for the current
-        axis).
-    - grid: bool, default False
-        Enables/disables the grid on the plot.
-    - labelsize: float, default fontsize
-        Sets the labels fontsize (which is the same for both labels).
-        The default value corresponds to the value of the keyword 'fontsize'.
-    - levels: np.ndarray
-        The levels of the contour lines.
-    - minorticks: str, default None
-        If not None enables the minor ticks on the plot (for both grid axes).
-    - sharex: Matplotlib axis | False, default False
-        Shares the x-axis with another axis.
-    - sharey: Matplotlib axis | False, default False
-        Shares the y-axis with another axis.
-    - ticksdir: {'in', 'out'}, default 'in'
-        Sets the ticks direction. The default option is 'in'.
-    - tickssize: float | bool, default True
-        Sets the ticks fontsize (which is the same for both grid axes).
-        The default value corresponds to the value of the keyword 'fontsize'.
-    - title: str, default None
-        Places the title of the plot on top of it.
-    - titlepad: float, default 8.0
-        Sets the distance between the title and the top of the plot
-    - titlesize: float, default fontsize
-        Sets the title fontsize. The default value corresponds to the value
[…]
