### 2025-07-02 a03fc15
removed windows 2019 tests
