+++ Src/pyPLUTO/configure.py
@@ -9,7 +9,7 @@ class Configure:
     """Class to handle the init tools for pyPLUTO. Tools include finding the
     session, setting up the handlers, and other initialization tasks."""
 
-    _greeted = False
+    greeted = False
 
     def __init__(
         self, colorerr: bool = True, colorwarn: bool = True, greet: bool = True
@@ -20,7 +20,7 @@ class Configure:
         self.colorwarn: bool = colorwarn
         self.session: str = self._find_session()
         self._setup_handlers(colorwarn, colorerr)
-        if greet and not Configure._greeted:
+        if greet and not Configure.greeted:
             print(f"PyPLUTO version: {self.version}   session: {self.session}")
             Configure._greeted = True
 
@@ -107,11 +107,36 @@ class Configure:
         _file: str | None = None,
         _line: str | None = None,
     ) -> str:
-        """Color the warnings in yellow."""
+        """Method to color the warnings in yellow.
+
+        Parameters
+        ----------
+        - message: str
+            The warning message.
+        - category: type[Warning]
+            The category of the warning.
+        - filename: str
+            The name of the file where the warning occurred.
+        - lineno: int
+            The line number where the warning occurred.
+        - _file: str | None, optional
+            Not used, kept for compatibility.
+        - _line: str | None, optional
+            Not used, kept for compatibility.
+
+        Returns
+        -------
+        - str
+            The formatted warning message with color codes.
+
+        """
+        # Format the warning message with color codes
         message = (
             f"\33[33m{category.__name__}: {message}"
             f"[{filename}:{lineno}]\33[0m\n"
         )
+
+        # Return the colored message
         return message
 
     def color_error(
@@ -120,16 +145,50 @@ class Configure:
         value: BaseException | None,
         tb: TracebackType | None,
     ) -> None:
-        """Color the errors in red."""
+        """Method to color the errors in red.
+
+        Parameters
+        ----------
+        - _type: type[BaseException] | None
+            The type of the exception.
+        - value: BaseException | None
+            The exception instance.
+        - tb: TracebackType | None
+            The traceback object.
+
+        Returns
+        -------
+        - None
+
+        """
+        # Traces the error and writes it in red
         traceback_str = "".join(traceback.format_tb(tb))
         sys.stderr.write(f"\033[91m{traceback_str}\033[0m")
         sys.stderr.write(f"\33[31m{value}\33[0m\n")  # Red color for errors
 
+        # End of the function
+
     def _setup_handlers(
         self, colorwarn: bool = True, colorerr: bool = True
     ) -> None:
-        """Set up the handlers for the warnings and errors."""
+        """Set up the handlers for the warnings and errors.
+
+        Parameters
+        ----------
+        - colorwarn: bool, default True
+            If True, color the warnings in yellow.
+        - colorerr: bool, default True
+            If True, color the errors in red.
+
+        Returns
+        -------
+        - None
+
+        """
+        # Set up the "always" filter for warnings
         warnings.simplefilter("always")
+
+        # Set up the handlers for warnings and errors
         if colorwarn:
             warnings.formatwarning = self.color_warning  # type: ignore
         if colorerr:
