+++ .github/workflows/test_windows.yml
@@ -12,7 +12,6 @@ jobs:
     strategy:
       matrix:
         os:
-          - windows-2019
           - windows-2022
           - windows-2025
         python-version: ["3.10", "3.11", "3.12", "3.13"]
