### 2025-07-02 612d98f
fixed typo in configure
