### 2025-07-02 #8448 @GiMattia
Hi,
Yes, I managed to finish what I intended to do and update the user guide correspondingly. I just merged the changes into the main branch, so from my side, we are good to go.
