### 2025-07-14 #8448 @tnehezd
> [@tnehezd](https://github.com/tnehezd), [@FaustinCarter](https://github.com/FaustinCarter) – I think we're good for you both to be picking up your reviews again here.

Sure, I was going through the package last week. The main points are fine and easy to follow, but I have a remark. Maybe it's just me, who never used PLUTO before, but it would be useful to more details in the documentation about the file types (like vtk, dbl, etc...) --- I guess it's obvious for someone familiar with PLUTO but for beginners, it would be a great help to understand. :)  @GiMattia
