### 2025-07-20 4f9935d
added check on docstrings
