### 2025-07-20 ea9664a
correct badge mypy
