+++ Utils/rewrite_docstrings.py
@@ -1,0 +1,7 @@
+# rewrite_docstrings.py
+def main():
+    print("🛠️  Rewriter not implemented yet — coming soon.")
+
+
+if __name__ == "__main__":
+    main()
