### 2025-07-24 #8448 @tnehezd
Hi @GiMattia ,

Great, thanks!

I've successfully completed the installation, and I've checked the well-known examples like Orszag-Tang and planet-disk, along with some of the built-in test cases. I've also checket the GUI version, and of course, read the paper and went through the checklist.

My only problem was with the input file descriptions, which I've now updated. :)

You've put together clear examples and a nice GUI. From my side, everything looks good.

Do I have any other tasks, @arfon?
