### 2025-07-26 #8448 @FaustinCarter
# Code/Paper Review

This review covers both the manuscript and the codebase. I did not open issues in the codebase, but if the author (or editors) would like me to copy the recommendations brought up in the code-review section to the Github repo, I'm happy to do that. I think the changes are small enough that they don't need multiple issues to fix.

Once the author has made the appropriate changes to both the paper and the source code (or argued why they are unnecessary to my satisfaction) I will happily check off the remaining boxes.

## Manuscript

Overall the manuscript explains what and why PyPLUTO is useful and worthy of publication in JOSS. I believe this pacakge is a good candidate and that this paper/codebase should be published in JOSS with minor corrections.

### Recommendations
- Please add units to the figure axis labels.
- Some line-by-line edits follow (all optional, but I think they will make the paper better).

---

- **Line 1**:
  - Capitalize key words in the title: `PyPLUTO: A Data Analysis Python Package for the PLUTO Code`.

- **Line 16**:
  - Change “MagnetoHydroDynamics” → **"magnetohydrodynamics" (MHD)** (lowercase, not a proper noun).

- **Line 21**:
  - “a general analytical solution is precluded” → **"a general analytical solution is not possible"**.

- **Line 23**:
  - “leans on dedicated software” → **"relies on dedicated software tools"**.

- **Line 25**:
  - “a more versatile and built-in tool” → **"a more versatile and integrated solution"**.

- **Lines 30–32**:
  - Avoid nested parentheses:
    ```
    such as cosmic rays (Mignone et al., 2018), Lagrangian (Vaidya et al., 2018), or dust (Mignone et al., 2019)
    ```
    → **Use inline citations or brackets instead**:
    ```
    such as cosmic rays [Mignone et al., 2018], Lagrangian particles [Vaidya et al., 2018], or dust particles [Mignone et al., 2019]
    ```

- **Line 44**:
  - “a complete rewrite of the original version” → **“a complete rewrite of the original”**

- **Line 46**:
  - Begin with “In addition to” rather than “Despite” to match tone:
    → **“In addition to its enhanced flexibility…”**

- **Line 55**:
  - “is made of mainly 3 classes” → **“primarily consists of three main classes”**

- **Lines 81–85**:
  - Clarify spatial references and punctuation in bullets:
    - “top right axis corner” → **“upper-right subplot”**
    - “top left axis corner” → **“upper-left subplot”**
    - Add full stops to all bullet points.

- **Line 114**:
  - “publication- quality figures” → **“publication-quality figures”**

- **Line 122**:
  - “including the possibility of visualizing particle distributions and trajectories dynamically” →
    **“including dynamic visualization of particle distributions and trajectories”**

- **Line 138**:
  - Long list of testers—recommend separating names with **semicolons** for clarity.

---

## Documentation

The documentation is visually appealing and mostly covers the basics. I find the usage documentation sparse, but the "Examples" section does an acceptable job of demonstrating core functionality. The keyword arguments that comprise the public API are also well documented.

Two major areas are missing: **developer documentation** and **GUI documentation**. Given that the GUI is a key functionality claim in the paper, it requires documentation that is commensurate with its role. Since this is a submission to JOSS, the documentation should also include developer contribution expectations and a rough overview of how the codebase works.

There is a folder in the repository named `Utils`, but it contains no documentation at all.

### Recommendations

- Provide clear documentation for launching and using the GUI. Ideally, this would be handled via an entry-point declaration in `pyproject.toml` (see "Packaging" section).
- Add developer-centric documentation covering, at a minimum:
  - A link to the source code on GitHub
  - Instructions for creating a development environment
  - How to run tests
  - Pre-commit expectations (formatting, style guide, etc.)
  - Notes on codebase architecture, especially considering the non-traditional structure that could trip up typical scientific coders
  - Intended use of the utilities in the `Utils` folder
- Run all documentation through an LLM or grammar checker to clean up grammar, punctuation, and clarity

---

## Examples

The examples ran smoothly after installation, except that they only work when launched from the `Examples` directory due to hardcoded file paths.

### Recommendations

- Update the data file paths to be coded relative to `__file__`, so the examples function correctly regardless of the working directory.

---

## GUI

The GUI is confusing and undocumented. It is not obvious how to load files or what types of files should be loaded. I eventually discovered that selecting any file and clicking "Reload Folder" produced plots that were much prettier than if I just selected a single file and clicked Plot.

Issues encountered:

- "Update Plot" throws a `'Image' object has no attribute '_set_cscale'` error every time.
- Plotting `y` vs. `x` instead of `x` vs. `y`
- Trying to plot anything against `z`
- Checking the "Transpose" box
- Strangely, checking "Transpose" *and* plotting `y vs. x` works!
- I stopped testing after this

### Recommendations

- The GUI needs documentation. A dedicated page should describe the function of each control and ideally include a short tutorial or use-case walkthrough.

---

## Packaging

The codebase contains a library (`pyPLUTO`) and a GUI that uses this library to provide access for non-developers. However, the GUI code is not included in the main package, and there are no instructions for running it. Since the GUI is a primary feature in the JOSS paper, both of these issues must be addressed.

There are two separate methods provided for building and installing the library. The preferred, modern method—using `pyproject.toml` and `pip`—is present and should be the only supported method. The alternative method uses a `setup.py` located in a non-standard path, which invites potential divergence between the two approaches.

The primary distribution channel is `https://plutocode.ph.unito.it`, which uses an untrusted SSL certificate. This makes it unusable for many users. The fallback is GitHub, which is acceptable but not ideal. The best practice would be to distribute the package on the Python Package Index (PyPI).

### Recommendations

- Unify the installation method by removing the `setup.py` file, or if backward compatibility is essential, move `setup.py` to the root and have it pull from `pyproject.toml`, as documented here:
  https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
- Move the GUI code into `src/pypluto` as a subpackage and add a command-line entry point in `pyproject.toml` so the GUI can be launched with a single command.
- Update the documentation accordingly and include GUI launch instructions.
- Optional: Upload the package to PyPI so users can install it via `pip install pypluto`.

---

## Codebase

The codebase is organized into a `src` folder containing the `pypluto` package, a GUI folder, a `Utilities` folder, and a `Tests` folder. The GUI folder should be relocated into the main package (see Packaging section). The `pypluto` package itself consists of many modules arranged in a flat hierarchy.

The primary interface is the `Image` class. Two secondary loader classes read data from disk and pass it to `Image`. The rest of the modules fall into three categories:

1. Modules that offer mixin-style functionality (imported directly into other classes)
2. Modules that define structural helpers (e.g., decorators, dataclasses)
3. "Manager" classes that extend `Image` functionality via a convoluted delegation system, hijacking `__getattr__` and `__setattr__`.

While the code works, the architecture relies on non-standard, overly complex patterns that make static analysis difficult, obfuscate method origins, and degrade IDE features such as autocomplete.

### Recommendations

- Organize the codebase into folders:
  - Place all manager modules into a `managers` folder
  - Place mixin-style helpers into a `mixins` folder
  - Place decorators and utilities into a `utils` folder
- Leave core modules (`image.py`, `imagestate.py`, `configure.py`, `load.py`, `loadpart.py`) at the top level for clarity

Minor improvements (many of which would be flagged by `ruff`):

- Avoid using `vars` or `dir` as variable names (they shadow Python built-ins)
- Replace `if key in some_dict.keys()` with `if key in some_dict`
- Ensure every module has a module-level docstring
- Prefer dict literals (`{}`) over `dict([...])` syntax
- Avoid variable shadowing of function arguments (possible bug due to this in `amr._inspect_hdf5`)
- Don't use mutable default arguments (e.g., lists, dicts; see `amr.oplotbox` as an example)
- Use `raise ... from ...` in `try/except` blocks to preserve tracebacks
- Replace `if not x == y` with `if x != y`
- Avoid use of `exec` on unsanitized input (security risk)

---

## Optional Pattern Critiques

These are subjective observations intended to improve long-term maintainability and developer experience. The code works—but adopting more conventional approaches could help reduce complexity.

### Pattern 1: Hijacking `__getattr__` and `__setattr__` for delegation

This pattern makes the `Image` class act as a façade for many "manager" classes via a `Mediator`. While clever, this design:

- Obscures the origin of methods and attributes
- Prevents static analysis tools and IDEs from resolving symbols
- Introduces fragility due to potential method name conflicts

**Recommendations**:

- Eliminate the `@delegate` decorator and `Mediator` class
- Use mixin-style classes or explicit property decorators to expose manager functionality in `Image`
- Delegate access to `ImageState` attributes using properties, rather than intercepting attribute access with `__getattr__`

### Pattern 2: Importing functions directly into class definitions

This approach provides shared functionality but deviates from common practice.

**Recommendations**:

- Refactor shared functions into mixin classes and use multiple inheritance
- Alternatively, extract shared functionality into separate helper classes and use composition

### Pattern 3: Overuse of `**kwargs`

The project uses a sprawling `kwargs` structure that necessitates a decorator for keeping track of what is and isn't allowed. This approach complicates type checking and obscures the API.

**Recommendations**:

- Refactor to use explicit arguments with default values where possible
- Use `mpl_kwargs` or similar only when forwarding options to third-party libraries (e.g., matplotlib)
- Avoid generic `**kwargs` in public APIs unless absolutely necessary
