### 2025-07-26 #8448 @tnehezd
Dear @GiMattia,

Here is the summary of my review about PyPLUTO

**JOSS Review Summary and Recommendations**

Your paper and the corresponding Git repo indicates that the submission meets the standards of JOSS, demonstrating a high-quality contribution to scientific software.

*Submitted paper and the statement of need of the PyPLUTO package*

The paper presents the authors and their affiliations and provides a brief summary of PyPLUTO. It highlights the clear statement of need, emphasizing that while other tools like plutoplot and the original, 13-year-old PyPLUTO version exist, they "may not support all data formats or offer integration for tasks like data manipulation and advanced plotting." PyPLUTO v1.0.0 addresses this by providing user-friendly methods for post-processing the diverse file outputs from the PLUTO code, enabling an efficient handling of large datasets for analyzing and visualizing the output data of PLUTO. The paper also mentions ongoing projects, for example, those exploring physical processes in protoplanetary disks, including the Magnetorotational Instability (MRI) and dust acceleration.

*Formal Compliance*

The paper includes an active link to the GitHub repository of the package, and a BSD 3-Clause license file is correctly included in the repository, fully complying with OSI approval criteria.

*Substantial Scholarly Effort & Installation*

The software represents substantial scholarly effort, The GitHub repository is active with the contributions of multiple authors and the commit history spans over two years.

The installation process was straightforward and clear, allowing for an easy setup. The package includes example usage, test cases and the GUI. I successfully ran and interacted with these parts, using several provided examples. While I didn't test with “real” PLUTO generated data (due to unfamiliarity with the PLUTO code itself), the provided examples and tests functioned as expected, demonstrating the software's core capabilities on post-processing and visualizing data that is substantial for scientific analysis.

*Code Quality & Community Engagement*

The project is hosted on GitHub. As an astronomer/astrophysicist, rather than a dedicated programmer, I found the code's structure and documentation clear and accessible. For further development, community contribution guidelines are provided directly within the Git repository, including details on pre-commit hooks and deeper checks.

**Recommended Revisions and Suggestions**

I have a few minor revisions and suggestions to further enhance its quality. These are primarily related to typographical corrections, clarity, and overall user-friendliness, rather than core functionality or code structure.

_1. Typo and Missing Reference_: Please address the missing space after the reference parenthesis in line 107 and the missing reference indicated by a '?' in line 108 of the paper. These are minor revisions that are essential for publication.
_2. PyPI Installation Instructions (Suggestion) and Naming Conflict_: I read in your previously comment that you are actively working to upload the package to PyPI, and facing naming issues. Once the PyPI naming situation is resolved, please ensure all documentation, including the README and any in-code references, is updated to reflect the final, chosen package name.
_3. Clarity on Repository Links (Suggestion for Paper Text)_: While the current repository link is present on the first page of the paper, I suggest adding a clarifying sentence in the paper text where the original version of PyPLUTO is referenced (e.g., "While a previous version of PyPLUTO is available [link to old repo], readers should note that this submission, PyPLUTO v1.0.0, is a complete rewrite hosted at a new repository [link to current repo]."). This would prevent potential confusion for readers navigating between the current and the old, 13 year old versions of the software.
_4. Contact Information for Contributors (Suggestion for Documentation)_: The description in the repository clearly includes the contribution rules, it would be beneficial to include some kind of contact information (e.g., an email address, a preferred communication channel) within the documentation for those wishing to ask questions or make enquiries beyond pull requests.

*Additional suggestion for the future work*
_Jupyter/IPython Notebook Examples (Suggestion)_: The existing examples are really helpful for understanding the usage of the package. However, including Jupyter or IPython Notebooks into the repository would be significantly useful for users, especially for those who are less familiar with the PLUTO package.  This is a suggestion for future development and user-friendliness, it is not a requirement from my end for this submission to be accepted.

Addressing these points will put your submission in an excellent position for publication in JOSS.
