+++ Paper/paper.bib
@@ -304,7 +304,7 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{Mattia_etal_2023,
+@ARTICLE{Mattiaetal2023,
        author = {{Mattia}, G. and {Del Zanna}, L. and {Bugli}, M. and {Pavan}, A. and {Ciolfi}, R. and {Bodo}, G. and {Mignone}, A.},
         title = "{Resistive relativistic MHD simulations of astrophysical jets}",
       journal = {\aap},
