+++ Examples/test07_khi.py
@@ -21,15 +21,18 @@ parameter in the pluto.ini should be changed to 0.05.
 # Loading the relevant packages
 import pyPLUTO
 
-# Loading the data into a pload object D
-Data = pyPLUTO.Load(path="Test_Problems/RMHD/KH")
+# Set the relative path to the data folder
+data_path = pyPLUTO.find_example("RMHD/KH")
+
+# Load data
+Data = pyPLUTO.Load(path=data_path)
 
 # Creating the image and the subplot axes (to have two secondary plots)
 Image = pyPLUTO.Image(
     figsize=[10.5, 10],
     suptitle="Test 07 - RMHD Kelvin-Helmholtz instability test",
 )
-Image.create_axes(ncol=2, wratio=[1, 0.1], wspace=[0.003], right=0.55)
+Image.create_axes(right=0.55)
 Image.create_axes(nrow=2, hspace=[0.003], left=0.67)
 
 # Plotting the data
@@ -43,9 +46,10 @@ Image.display(
     shading="gouraud",
     xtitle=r"$x$",
     ytitle=r"$y$",
+    cpos="right",
 )
 
-Image.colorbar(axs=0, cax=1)
+# Image.colorbar(axs=0, cax=1)
 
 # Find and plot the field lines
 lines = Data.find_fieldlines(
@@ -67,14 +71,14 @@ for _, line in enumerate(lines):
 analysis = Data.read_file("kh.dat")
 
 # Add text in the axes
-Image.text(r"$\langle v_y^2\rangle$", ax=2, x=0.05)
-Image.text(r"$v_{y, MAX}^2$", ax=3, x=0.05)
+Image.text(r"$\langle v_y^2\rangle$", ax=1, x=0.05)
+Image.text(r"$v_{y, MAX}^2$", ax=2, x=0.05)
 
 # Plot the velocity from the kh.dat file.
 Image.plot(
     analysis["time"],
     analysis["vy2"],
-    ax=2,
+    ax=1,
     c="k",
     yscale="log",
     xtickslabels=None,
@@ -82,12 +86,12 @@ Image.plot(
 Image.plot(
     analysis["time"],
     analysis["maxvy"],
-    ax=3,
+    ax=2,
     c="k",
     yscale="log",
     xtitle=r"$t$",
 )
 
 # Saving the image and showing the plots
-Image.savefig("test07_khi.png")
+Image.savefig("test07_khi.png", script_relative=True)
 pyPLUTO.show()
