### 2025-07-29 cbd13fc
added guiplot documentatino file
