### 2025-07-29 b5e050f
fixed GUI bugs
