### 2025-07-29 1c5cbaa
changed CI path tests
