+++ Docs/Figures/test_gui_sod.py
@@ -1,0 +1,237 @@
+import re
+import shutil
+import subprocess
+import sys
+import time
+from pathlib import Path
+
+from PyQt6.QtCore import QTimer
+from PyQt6.QtWidgets import QApplication
+
+import pyPLUTO
+from pyPLUTO.gui.main_window import PyPLUTOApp
+
+
+def get_window_geometry(window_title="PyPLUTO GUI"):
+    # Find window ID by title
+    win_id = (
+        subprocess.check_output(["xdotool", "search", "--name", window_title])
+        .decode()
+        .strip()
+        .split("\n")[0]
+    )
+
+    # Get window geometry
+    output = subprocess.check_output(["xwininfo", "-id", win_id]).decode()
+
+    x = int(re.search(r"Absolute upper-left X:\s+(\d+)", output).group(1))
+    y = int(re.search(r"Absolute upper-left Y:\s+(\d+)", output).group(1))
+    width = int(re.search(r"Width:\s+(\d+)", output).group(1))
+    height = int(re.search(r"Height:\s+(\d+)", output).group(1))
+
+    return x, y, width, height
+
+
+def record_screen_start(filename="pluto_sod_full.mp4", display=":0.0", fps=25):
+    x, y, w, h = get_window_geometry("PyPLUTO")
+
+    screen_width = 1920
+    screen_height = 1080
+
+    # Adjust width and x if necessary
+    if x + w > screen_width:
+        w = screen_width - x
+    # Adjust height and y if necessary
+    if y + h > screen_height:
+        h = screen_height - y
+
+    # Ensure even dimensions for x264
+    w -= w % 2
+    h -= h % 2
+    cmd = [
+        "ffmpeg",
+        "-y",
+        "-f",
+        "x11grab",
+        "-framerate",
+        str(fps),
+        "-video_size",
+        f"{w}x{h}",
+        "-i",
+        f"{display}+{x},{y}",
+        "-vcodec",
+        "libx264",
+        "-preset",
+        "ultrafast",
+        "-pix_fmt",
+        "yuv420p",
+        filename,
+    ]
+    print(f"Recording {w}x{h} at ({x},{y}) → {filename}")
+    return subprocess.Popen(cmd, stdin=subprocess.PIPE)
+
+
+def record_screen_stop(proc):
+    if proc.poll() is not None:
+        print("ffmpeg already exited.")
+        return
+
+    print("Stopping ffmpeg recording gracefully...")
+    proc.stdin.write(b"q\n")
+    proc.stdin.flush()
+    try:
+        proc.wait(timeout=10)
+        print("ffmpeg exited cleanly.")
+    except subprocess.TimeoutExpired:
+        print("Timeout — killing ffmpeg.")
+        proc.kill()
+
+
+def postprocess_video():
+    cmd = [
+        "ffmpeg",
+        "-y",  # Overwrite without asking
+        "-ss",
+        "0.25",
+        "-i",
+        "pluto_sod_full.mp4",
+        "-t",
+        "18",
+        "-vf",
+        "crop=1180:750:40:47",
+        "-c:a",
+        "copy",
+        "pluto_sod_gui.mp4",
+    ]
+    print("Running postprocess crop + trim...")
+    subprocess.run(cmd, check=True)
+
+
+def copy_to_static(
+    source_video="pluto_sod_gui.mp4", static_dir="../source/_static"
+):
+    src = Path(source_video)
+    dst_dir = Path(static_dir).resolve()
+    dst_dir.mkdir(parents=True, exist_ok=True)
+    dst = dst_dir / src.name
+
+    shutil.copy2(src, dst)
+    print(f"✅ Copied {src} → {dst}")
+
+
+def automate_gui():
[…]
