### 2025-07-30 09c5ef1
try fix for self.figure
