### 2025-07-30 0cea7fc
fixed colorbar for windows?
