### 2025-07-30 3d6f600
failed idea, reverting
