### 2025-07-30 7be4c74
self.state.fig in colorbar
