### 2025-07-30 325c232
"Untrack all files listed in .gitignore"
