+++ coverage_tests.py
@@ -1,197 +1,0 @@
-import os
-import re
-import subprocess
-import sys
-
-cov_errs = 50
-cov_warn = 75
-
-RED = "\033[31m"
-YELLOW = "\033[33m"
-GREEN = "\033[32m"
-RESET = "\033[0m"
-
-globscore = 0
-
-commands = [
-    {
-        "name": "pytest",
-        "cmd": ["pytest", "--cov=pyPLUTO", "--cov-report=term-missing"],
-        "output": "Utils/pytest_output.txt",
-    },
-    {
-        "name": "mypy",
-        "cmd": ["mypy", "--strict", "@mypy_files.txt"],
-        "output": "Utils/mypy_output.txt",
-    },
-    {
-        "name": "pylint",
-        "cmd": None,  # We'll fill this dynamically based on mypy_files.txt
-        "output": "Utils/pylint_output.txt",
-    },
-    {
-        "name": "interrogate",
-        "cmd": ["interrogate", "Src/pyPLUTO/"],
-        "output": "Utils/interrogate_output.txt",
-    },
-]
-
-
-def read_mypy_files_list(file_path: str = "mypy_files.txt") -> list[str]:
-    with open(file_path) as f:
-        return [
-            line.strip()
-            for line in f
-            if line.strip() and not line.startswith("#")
-        ]
-
-
-def run_command(cmd: list[str], output_file: str):
-    try:
-        result = subprocess.run(
-            cmd,
-            text=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            check=False,  # Let it continue even on failures
-        )
-        with open(output_file, "w") as f:
-            f.write(result.stdout)
-    except Exception as e:
-        with open(output_file, "w") as f:
-            f.write(f"Error running command: {' '.join(cmd)}\n{e}")
-
-
-def colorize_test(score: float, summary: str) -> str:
-    global globscore
-
-    if score < cov_errs:  # 1 error every 2 files or more
-        color = RED
-        globscore += 0
-    elif score < cov_warn:  # 1 error every 4 files or more
-        color = YELLOW
-        globscore += 1
-    else:
-        color = GREEN
-        globscore += 2
-
-    return f"{color}{summary}{RESET}"
-
-
-def extract_score(tool_name: str, output_file: str) -> str:
-    with open(output_file) as f:
-        content = f.read()
-
-    if tool_name == "pytest":
-        errors_match = re.search(r"=+\s*(\d+)\s+failed? in [\d\.]+s", content)
-        errors_count = int(errors_match.group(1)) if errors_match else 0
-
-        warnings_match = re.search(
-            r"=+\s*\d+\s+passed,\s*(\d+)\s+warnings? in [\d\.]+s", content
-        )
-        warnings_count = int(warnings_match.group(1)) if warnings_match else 0
-
-        # Extract coverage percentage from TOTAL line
-        coverage_match = re.search(
-            r"TOTAL\s+\d+\s+\d+\s+\d+\s+\d+\s+(\d+)%", content
-        )
-        coverage = float(coverage_match.group(1)) if coverage_match else 0
-
-        passed_match = re.search(
-            r"=+\s*(\d+\s+passed[^i]*)\s+in\s+[\d\.]+s", content
-        )
-        if passed_match:
-            passed_text = passed_match.group(1).strip()
-        elif errors_match:
-            passed_text = f"{errors_count} errors, "
-        else:
-            passed_text = f"{warnings_count} warnings, "
-
-        summary = f"{passed_text} (coverage {coverage}%)"
-
-        coverage = (
-            cov_errs - 1
-            if "warnings" in passed_text or "errors" in passed_text
-            else coverage
-        )
-
[…]
