+++ Utils/scan_exposed.py
@@ -1,326 +1,0 @@
-import ast
-import json
-import os
-import re
-from collections import defaultdict
-from pathlib import Path
-
-# --- CONFIGURATION ---
-UTILS_DIR = Path(__file__).parent.resolve()
-SRC_DIR = (UTILS_DIR / "../Src/pyPLUTO").resolve()
-JSON_FILE = UTILS_DIR / "exposed_kwargs.json"
-
-
-def extract_doc_params(docstring: str) -> tuple[dict[str, str], int]:
-    if not docstring:
-        return {}, 0
-
-    raw_lines = docstring.splitlines()
-
-    # Skip first line when checking indentation
-    indent_level = 0
-    for line in raw_lines[1:]:
-        if line.strip():
-            indent_level = len(line) - len(line.lstrip(" "))
-            break
-
-    lines = [line.rstrip() for line in raw_lines]
-    param_docs = {}
-    in_params = False
-    i = 0
-
-    while i < len(lines):
-        if lines[i].strip().lower() == "parameters":
-            i += 1
-            while i < len(lines) and re.match(r"^\s*-{3,}", lines[i]):
-                i += 1
-            in_params = True
-            break
-        i += 1
-
-    current_param = None
-    current_desc = []
-
-    def flush():
-        if current_param and current_desc:
-            first, *rest = current_desc
-            param_docs[current_param] = first + (
-                "\n" + " ".join(rest) if rest else ""
-            )
-
-    while in_params and i < len(lines):
-        line = lines[i]
-
-        # STOP if reaching next section like ---- or Examples
-        if re.match(r"^\s*-{3,}\s*$", line) or line.strip().lower().startswith(
-            "examples"
-        ):
-            flush()
-            break
-
-        param_match = re.match(
-            r"^\s*-?\s*([a-zA-Z_]\w*)[^\:]*\s*:\s*(.*)", line
-        )
-        if param_match:
-            flush()
-            current_param = param_match.group(1).strip()
-            current_desc = [param_match.group(2).strip()]
-        elif line.startswith(" ") or line.startswith("\t"):
-            current_desc.append(line.strip())
-        elif line.strip() == "":
-            pass
-        else:
-            # Any unrecognized line = end of parameter block
-            flush()
-            break
-        i += 1
-
-    flush()
-    return param_docs, indent_level
-
-
-def find_python_files(root: Path):
-    for dirpath, _, files in os.walk(root):
-        for fname in files:
-            if fname.endswith(".py"):
-                yield Path(dirpath) / fname
-
-
-def get_class_defs(tree):
-    return {
-        node.name: node for node in tree.body if isinstance(node, ast.ClassDef)
-    }
-
-
-def extract_exposed_methods(class_node):
-    for stmt in class_node.body:
-        if isinstance(stmt, ast.Assign):
-            for target in stmt.targets:
-                if (
-                    isinstance(target, ast.Name)
-                    and target.id == "exposed_methods"
-                ):
-                    v = stmt.value
-                    if isinstance(v, (ast.Tuple, ast.List)):
-                        return [
-                            elt.value
-                            for elt in v.elts
-                            if isinstance(elt, ast.Constant)
-                            and isinstance(elt.value, str)
-                        ]
-    return []
-
-
-def get_attr_class_map(class_node):
-    mapping = {}
-    for node in class_node.body:
-        if isinstance(node, ast.FunctionDef) and node.name == "__init__":
-            for stmt in ast.walk(node):
-                if (
-                    isinstance(stmt, ast.Assign)
-                    and isinstance(stmt.targets[0], ast.Attribute)
-                    and isinstance(stmt.targets[0].value, ast.Name)
[…]
