+++ ruff_errors.txt
@@ -1,1 +1,0 @@
-Fixed 54 errors (188 additional fixes available with `--unsafe-fixes`).
