### 2025-07-31 bb0e1bb
updated automated tests
