+++ .github/workflows/test_windows.yml
@@ -2,7 +2,7 @@ name: Tests on Windows
 
 on:
   push:
-    branches: [ master ]
+    #branches: [ master ]
   pull_request:
 
 jobs:
@@ -40,11 +40,6 @@ jobs:
         run: |
           pip uninstall -y pyPLUTO || true
           pip install -e .
-
-      - name: Show import path (debug)
-        shell: pwsh
-        run: |
-          python -c "import pyPLUTO; print('pyPLUTO loaded from:', pyPLUTO.__file__)"
 
       - name: Run Tests
         shell: pwsh
