+++ .github/workflows/test_macos.yml
@@ -2,8 +2,7 @@ name: Tests on MacOS
 
 on:
   push:
-    branches:
-      - master
+    #branches: [ master ]
   pull_request:
 
 jobs:
@@ -21,6 +20,13 @@ jobs:
       - name: Checkout repository
         uses: actions/checkout@v4
 
+      - name: Clean untracked files & bytecode
+        shell: pwsh
+        run: |
+          git clean -xfd
+          Get-ChildItem -Path . -Include *.pyc -Recurse | Remove-Item -Force -ErrorAction SilentlyContinue
+          Get-ChildItem -Path . -Include __pycache__ -Recurse | Remove-Item -Force -Recurse -ErrorAction SilentlyContinue
+
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
@@ -31,9 +37,11 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements_dev.txt
 
-      - name: Install pyPLUTO
+      - name: Install pyPLUTO in editable mode
         run: |
-          python -m pip install .
+          pip uninstall -y pyPLUTO || true
+          pip install -e .
+
 
       - name: Run Tests
         run: pytest -v
