### 2025-07-31 000aafc
changed self.state.ffig in image imxin
