### 2025-08-03 8acddb1
added commit manuscript action
