+++ .github/workflows/draft-pdf.yml
@@ -30,3 +30,10 @@ jobs:
           # PDF. Note, this should be the same directory as the input
           # paper.md
           path: Paper/paper.pdf
+
+      - name: Commit PDF to repository
+        uses: EndBug/add-and-commit@v9
+        with:
+          message: '(auto) Paper PDF Draft'
+          # This should be the path to the paper within your repo.
+          add: 'Paper/paper.pdf' # 'paper/*.pdf' to commit all PDFs in the paper directory
