+++ Paper/paper.bib
@@ -477,3 +477,73 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2024JCoPh.49912701B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@ARTICLE{MelonFuksman2025,
+       author = {{Melon Fuksman}, David and {Flock}, Mario and {Klahr}, Hubert and {Mattia}, Giancarlo and {Muley}, Dhruv},
+        title = "{Multidimensional half-moment multigroup radiative transfer. Improving moment-based thermal models of circumstellar disks}",
+      journal = {arXiv e-prints},
+     keywords = {Instrumentation and Methods for Astrophysics, Earth and Planetary Astrophysics, Astrophysics of Galaxies, Solar and Stellar Astrophysics},
+         year = 2025,
+        month = apr,
+          eid = {arXiv:2504.13999},
+        pages = {arXiv:2504.13999},
+          doi = {10.48550/arXiv.2504.13999},
+archivePrefix = {arXiv},
+       eprint = {2504.13999},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250413999M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Costa2025,
+       author = {{Costa}, A. and {Bodo}, G. and {Tavecchio}, F. and {Rossi}, P. and {Coppi}, P. and {Sciaccaluga}, A. and {Boula}, S.},
+        title = "{How do recollimation-induced instabilities shape the propagation of hydrodynamic relativistic jets?}",
+      journal = {arXiv e-prints},
+     keywords = {High Energy Astrophysical Phenomena},
+         year = 2025,
+        month = mar,
+          eid = {arXiv:2503.18602},
+        pages = {arXiv:2503.18602},
+          doi = {10.48550/arXiv.2503.18602},
+archivePrefix = {arXiv},
+       eprint = {2503.18602},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250318602C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Harer2025,
+       author = {{H{\"a}rer}, L. and {Vieu}, T. and {Reville}, B.},
+        title = "{Stellar-wind feedback and magnetic fields around young compact star clusters: 3D magnetohydrodynamics simulations}",
+      journal = {\aap},
+     keywords = {acceleration of particles, magnetohydrodynamics (MHD), stars: winds, outflows, ISM: bubbles, open clusters and associations: general, High Energy Astrophysical Phenomena, Astrophysics of Galaxies},
+         year = 2025,
+        month = jun,
+       volume = {698},
+          eid = {A6},
+        pages = {A6},
+          doi = {10.1051/0004-6361/202554057},
+archivePrefix = {arXiv},
+       eprint = {2503.19745},
+ primaryClass = {astro-ph.HE},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025A&A...698A...6H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Sciaccaluge2025,
+       author = {{Sciaccaluga}, A. and {Costa}, A. and {Tavecchio}, F. and {Bodo}, G. and {Coppi}, P. and {Boula}, S.},
+        title = "{The polarization of the synchrotron radiation from a recollimated jet: Application to high-energy BL Lacs}",
+      journal = {\aap},
[…]
