### 2025-08-27 442a753
added references paper + removed vale.ini
