+++ .vale.ini
@@ -1,5 +1,0 @@
-MinAlertLevel = suggestion
-StylesPath = .vale
-
-[*.{rst,md}]
-BasedOnStyles = Vale
