### 2025-08-27 ef288a2
fixed paper references
