+++ Paper/paper.bib
@@ -31,7 +31,7 @@
 }
 
 @ARTICLE{MelonFuksman2021,
-       author = {{Melon Fuksman}, Julio David and {Klahr}, Hubert and {Flock}, Mario and {Mignone}, Andrea},
+       author = {{Melon Fuksman}, D. and {Klahr}, H. and {Flock}, M. and {Mignone}, A.},
         title = "{A Two-moment Radiation Hydrodynamics Scheme Applicable to Simulations of Planet Formation in Circumstellar Disks}",
       journal = {\apj},
      keywords = {Radiative transfer, Protoplanetary disks, Planet formation, 1335, 1300, 1241, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
@@ -50,7 +50,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Muley2024a,
-       author = {{Muley}, Dhruv and {Melon Fuksman}, Julio David and {Klahr}, Hubert},
+       author = {{Muley}, D. and {Melon Fuksman}, D. and {Klahr}, H.},
         title = "{Three-temperature radiation hydrodynamics with PLUTO: Thermal and kinematic signatures of accreting protoplanets}",
       journal = {\aap},
      keywords = {methods: numerical, protoplanetary disks, planet-disk interactions, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
@@ -253,7 +253,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Vaidya2018,
-       author = {{Vaidya}, Bhargav and {Mignone}, Andrea and {Bodo}, Gianluigi and {Rossi}, Paola and {Massaglia}, Silvano},
+       author = {{Vaidya}, B. and {Mignone}, A. and {Bodo}, G. and {Rossi}, P. and {Massaglia}, S.},
         title = "{A Particle Module for the PLUTO Code. II. Hybrid Framework for Modeling Nonthermal Emission from Relativistic Magnetized Flows}",
       journal = {\apj},
      keywords = {acceleration of particles, methods: numerical, polarization, radiation mechanisms: non-thermal, relativistic processes, shock waves, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
@@ -370,7 +370,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{MelonFuksman2024a,
-       author = {{Melon Fuksman}, Julio David and {Flock}, Mario and {Klahr}, Hubert},
+       author = {{Melon Fuksman}, D. and {Flock}, M. and {Klahr}, H.},
         title = "{Vertical shear instability in two-moment radiation-hydrodynamical simulations of irradiated protoplanetary disks. I. Angular momentum transport and turbulent heating}",
       journal = {\aap},
      keywords = {hydrodynamics, instabilities, radiative transfer, methods: numerical, protoplanetary disks, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Solar and Stellar Astrophysics, Physics - Computational Physics},
@@ -388,7 +388,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{MelonFuksman2024b,
-       author = {{Melon Fuksman}, Julio David and {Flock}, Mario and {Klahr}, Hubert},
+       author = {{Melon Fuksman}, D. and {Flock}, M. and {Klahr}, H.},
[…]
