### 2025-08-27 #8448 @GiMattia
Dear all,

Sorry for the delay, but I was also on vacation.
After the referees ' comments, the other collaborators and I have finished revising the manuscript and the Python package.

------------------------------------------------------------

Here I report our more detailed answers for @FaustinCarter:

**Manuscript Recommendations**

- _Please add units to the figure axis labels._
- _Some line-by-line edits follow (all optional, but I think they will make the paper better)._

We have added the axes labels and corrected the manuscript (we just left the commas to separate the testers in the acknowledgments). We have also improved a few sentences in the manuscript (minor modifications).

**Documentation Recommendations**

- _Provide clear documentation for launching and using the GUI. Ideally, this would be handled via an entry-point declaration in pyproject.toml (see "Packaging" section)._
- _Add developer-centric documentation._
- _Run all documentation through an LLM or grammar checker to clean up grammar, punctuation, and clarity._

We added a "Contribute" section in the documentation, showing how to create a development environment, run tests, and use the precommit. The section also includes a small section on the codebase architecture and the style guide. We have removed the Utils folder from the repository since it is not ready for deployment.
We have added more detailed instructions on the GUI. Users can now install it as a subpackage through the pyproject.toml.
Finally, we have run the packages through a grammar check.

**Examples Recommendations**

- _Update the data file paths to be coded relative to `__file__`, so the examples function correctly regardless of the working directory._

We have added a function called "find_example" that looks for the Test_Problem directory and, if not found, checks the PLUTO/Test_Problem directory. This solution not only fixes the absolute Path issue but also allows us to run the examples from this repository and the gPLUTO repository.

**GUI Recommendations**

- _The GUI needs documentation. A dedicated page should describe the function of each control and ideally include a short tutorial or use-case walkthrough._

We have added a few pages in the documentation explaining every part of the GUI and included two video tutorials covering most of the GUI functions. We have also spotted and fixed a few bugs.

**Packaging Recommendations**
- _Unify the installation method by removing the setup.py file, or if backward compatibility is essential, move setup.py to the root and have it pull from pyproject.toml, as documented here:
https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html_
- _Move the GUI code into src/pypluto as a subpackage and add a command-line entry point in pyproject.toml so the GUI can be launched with a single command._
- _Update the documentation accordingly and include GUI launch instructions._
- _Optional: Upload the package to PyPI so users can install it via pip install pypluto._

We have removed setup.py from the repository and the Src folder (the code folder will now be pyPLUTO since there is no need for a subfolder). The GUI has been moved as a subpackage and can be launched through the command "pypluto-gui." We have updated the documentation accordingly.

**Codebase Recommendations**
- _Organize the codebase into folders._
- _Minor improvements (many of which would be flagged by ruff)._

We have reorganized the folder into:
- gui folder ---> GUI
- imagefuncs ---> image managers
- loadfuncs ---> functions to load, read, and write files
- toolsfuncs ---> functions to manipulate the data
- utils ---> decorators
- codes ---> in case someone wants to load another code with pyPLUTO

We also did a full ruff check.

**Pattern 1: Hijacking `__getattr__` and `__setattr__` for delegation**

We have created a Mixin class and used properties to avoid obscure calls and name conflicts.
Please let us know if this design is satisfactory. It can also be adapted to the other classes (see next point).

**Pattern 2: Importing functions directly into class definitions**

We know about this issue and will take care of it soon.

**Pattern 3: Overuse of `**kwargs`**

We understand this point, but since most of the kwargs are related to plotting customization, we will keep the `**kwargs` in the public methods for now.

------------------------------------------------------------

Here I report our more detailed answers for @tnehezd:

**Typo and Missing Reference**
_Please address the missing space after the reference parenthesis in line 107 and the missing reference indicated by a '?' in line 108 of the paper. These are minor revisions that are essential for publication._

We have fixed the missing reference and added a few more examples about PyPLUTO's usage that appeared after the submission.

**PyPI Installation Instructions (Suggestion) and Naming Conflict**
_I read in your previously comment that you are actively working to upload the package to PyPI, and facing naming issues. Once the PyPI naming situation is resolved, please ensure all documentation, including the README and any in-code references, is updated to reflect the final, chosen package name._

We have sent another email to PyPI admins and hope to receive an answer soon. Once the name PyPLUTO becomes available, we will upload it to PyPI and change the documentation and in-code references to avoid confusion among users.

**Clarity on Repository Links**
_While the current repository link is present on the first page of the paper, I suggest adding a clarifying sentence in the paper text where the original version of PyPLUTO is referenced (e.g., "While a previous version of PyPLUTO is available [link to old repo], readers should note that this submission, PyPLUTO v1.0.0, is a complete rewrite hosted at a new repository [link to current repo]."). This would prevent potential confusion for readers navigating between the current and the old, 13 year old versions of the software._

We have the sentence in the paper accordingly. Please let us know if you prefer this version.

**Contact Information for Contributors**
_The description in the repository clearly includes the contribution rules, it would be beneficial to include some kind of contact information (e.g., an email address, a preferred communication channel) within the documentation for those wishing to ask questions or make enquiries beyond pull requests._

We have added my e-mail address (mattia@mpia.de) to the documentation to provide a clear contact in case needed. We are also putting a lot of effort into interacting with potential users at conferences, meetings, etc., to better communicate with users and potential contributors.

**Jupyter/IPython Notebook Examples**
_The existing examples are really helpful for understanding the usage of the package. However, including Jupyter or IPython Notebooks into the repository would be significantly useful for users, especially for those who are less familiar with the PLUTO package. This is a suggestion for future development and user-friendliness, it is not a requirement from my end for this submission to be accepted._

We are planning to release more tests in the future, including Jupyter/IPython Notebook Examples. So far, we have just made sure that PyPLUTO works on notebooks, but in the near future, we will build more tests that can also be used as package tutorials.
