### 2025-08-27 #8448 @tnehezd
Dear @GiMattia,

I have read the revised manuscript and your response to my comments. I appreciate the effort you put into updating the repository as a part of your response to the revision.

The modifications to the manuscript have improved the quality of the paper. I accept your responses, and from my side, I am satisfied with the revised version.

It has been a pleasure to work with you on this. Your work is a valuable contribution to the field, and I wish you all the best with your future work on this project.

Sincerely,
Dóri

----

Note to the Editor (@arfon): I have completed my review of the revised manuscript. All of my concerns have been addressed, and I recommend that the paper be accepted for publication.
