### 2025-08-29 #8448 @GiMattia
Hi @tnehezd,

Thank you very much for your quick reply and kind words. It was a pleasure to work with you as well.

Sincerely,
Giancarlo (and collaborators)
