### 2025-09-02 fb50626
ruff check
