### 2025-09-03 a8f33f9
adding custom var to GUI
