### 2025-09-03 7c2f45a
fixed typos in paper
