+++ Paper/paper.md
@@ -78,8 +78,8 @@ In recent years, numerical simulations have become indispensable for addressing 
 
 # Statement of Need
 
-The PLUTO code [@PLUTO_2007] is a widely used, freely distributed computational fluid dynamics code designed to solve the classical and (special) relativistic MHD equations in different geometries and spatial dimensions. The original code is written in C (while the upcoming GPU version provides a full [C++ rewrite](https://plutocode.ph.unito.it/pluto-gpu.html)) and it contains several numerical methods adaptable to different contexts. Data post-processing is a crucial step in analyzing the results of any numerical simulation. Other packages addressing related needs (e.g., [plutoplot](https://github.com/Simske/plutoplot)), provide valuable functionality for working with PLUTO data, including loading and visualization. However, they may not support all data formats or offer integration for tasks like data manipulation and advanced plotting.
-In this work, we present PyPLUTO, a Python package designed to load, manipulate and visualize efficiently the output from the PLUTO code.
+The PLUTO code [@PLUTO_2007] is a widely used, freely distributed computational fluid dynamics code designed to solve the classical and (special) relativistic MHD equations in different geometries and spatial dimensions. The original code is written in C (while the upcoming GPU version provides a full [C++ rewrite](https://plutocode.ph.unito.it/pluto-gpu.html)) and it contains several numerical methods adaptable to different contexts. Data post-processing is a crucial step in analyzing the results of any numerical simulation. Other packages addressing related needs (e.g., [plutoplot](https://github.com/Simske/plutoplot)) provide valuable functionality for working with PLUTO data, including loading and visualization. However, they may not support all data formats or offer integration for tasks like data manipulation and advanced plotting.
+In this work, we present PyPLUTO, a Python package designed to load, manipulate, and visualize efficiently the output from the PLUTO code.
 While a previous version of PyPLUTO is [available](https://github.com/coolastro/pyPLUTO), the package presented here is a complete rewrite hosted at a [new repository](https://github.com/GiMattia/PyPLUTO).
 The package retains its core strengths while offering user-friendly methods for generating publication-quality plots with high customization. In addition to its enhanced flexibility, PyPLUTO offers strong computational efficiency, enabling the rapid handling of large datasets typical of state-of-the-art numerical simulations. Through this balance between customization, performance, and ease of use, PyPLUTO represents a key tool to effectively communicate scientific results while minimizing the effort required for post-processing.
 
@@ -105,18 +105,18 @@ This test simulates the interaction of a planet embedded in a disk [@Mignone_eta
 
 • The first zoom (upper-right subplot) shows an enlarged view of the density distribution using the same color map and logarithmic scale as the global plot.
 
-• The second zoom (upper-left subplot) highlights the changes in toroidal velocity due to the presence of the planet by employing a different color map (to enhance the sign change) and a linear color scale.
+• The second zoom (lower-left subplot) highlights the changes in toroidal velocity due to the presence of the planet by employing a different color map (to enhance the sign change) and a linear color scale.
 
 These zoomed views offer deeper insights into the physical processes at play and demonstrate the utility of PyPLUTO for analyzing complex astrophysical systems.
 
 ### Particles Accelerated near an X-point {#xpoint}
[…]
