### 2025-09-05 d6689df
added custom variables GUI
