+++ pyPLUTO/gui/main_window.py
@@ -1,4 +1,10 @@
-from PyQt6.QtWidgets import QHBoxLayout, QMainWindow, QVBoxLayout, QWidget
+from PyQt6.QtWidgets import (
+    QHBoxLayout,
+    QMainWindow,
+    QTextEdit,
+    QVBoxLayout,
+    QWidget,
+)
 
 from .globals import cmaps, cmaps_avail, format_avail, scales, vscales
 
@@ -121,8 +127,12 @@ class PyPLUTOApp(QMainWindow):
 
         self.add_line(button_layout)
 
-        self.add_label("Information:", button_layout, "info_label", 250, 170)
-        self.info_label.setWordWrap(True)  # type: ignore
+        info_box = QTextEdit()
+        info_box.setObjectName("info_label")
+        info_box.setReadOnly(True)
+        info_box.setFixedSize(370, 200)  # keep your original fixed size
+        button_layout.addWidget(info_box)
+        self.info_label = info_box
 
         self.add_line(button_layout)
 
