+++ pyPLUTO/gui/custom_var.py
@@ -1,4 +1,11 @@
-# pypluto_custom_var_singlefile.py
+"""A custom variable evaluator for PyPLUTO."""
+
+import os
+import re
+import tempfile
+
+import numexpr as ne
+import numpy as np
 from PyQt6.QtCore import Qt
 from PyQt6.QtWidgets import (
     QComboBox,
@@ -6,155 +13,344 @@ from PyQt6.QtWidgets import (
     QDialogButtonBox,
     QFormLayout,
     QLabel,
-    QLineEdit,
     QPlainTextEdit,
     QVBoxLayout,
 )
 
 SENTINEL = "Custom var..."
 
+# --- tiny normalizer ---------------------------------------------------------
+# Allow both D.something and bare names; allow np./numpy. prefixes.
+_NORM_PATTERNS = [
+    (re.compile(r"\bD\."), ""),  # D.Bx1 -> Bx1
+    (re.compile(r"\bnp\."), ""),  # np.sqrt -> sqrt
+    (re.compile(r"\bnumpy\."), ""),  # numpy.sqrt -> sqrt
+]
 
-# ---- Minimal dialog: name (1 line) + expression (multi-line) ----
+
+def _normalize_expr(expr: str) -> str:
+    s = expr.strip()
+    for pat, repl in _NORM_PATTERNS:
+        s = pat.sub(repl, s)
+    return s
+
+
+# --- evaluator ---------------------------------------------------------------
+def evaluate_custom_var(D, name: str, expr: str, *, assign: bool = True):
+    """
+    numexpr-only, memmap output for arrays, scalar stays scalar.
+    Accepts D.foo or foo; np.func/numpy.func or func directly.
+    """
+    expr = _normalize_expr(expr)
+
+    # Build locals from D (public, non-callable) + constants
+    local = {"pi": float(np.pi), "e": float(np.e)}
+    for k, v in D.__dict__.items():
+        if not k.startswith("_") and not callable(v):
+            local[k] = v
+
+    # Parse/compile (syntax check)
+    try:
+        compiled = ne.NumExpr(expr)
+    except Exception as e:
+        raise ValueError(f"compile error: {e}")
+
+    # Validate cheaply on only the used names (1-element views)
+    names = compiled.input_names
+    tiny = {}
+    for n in names:
+        v = local.get(n)
+        if v is None:
+            raise ValueError(f"unknown name: {n}")
+        tiny[n] = v.reshape(-1)[:1] if isinstance(v, np.ndarray) else v
+    try:
+        tiny_res = ne.evaluate(expr, local_dict=tiny, global_dict={})
+    except Exception as e:
+        raise ValueError(f"validation error: {e}")
+
+    # Infer shape from arrays actually referenced (broadcasting)
+    arrs = [local[n] for n in names if isinstance(local[n], np.ndarray)]
+    if not arrs:
+        # Scalar result
+        try:
+            res = ne.evaluate(expr, local_dict=local, global_dict={})
+        except Exception as e:
+            raise ValueError(f"evaluate error: {e}")
+        if isinstance(res, np.ndarray) and res.shape == ():
+            res = res.item()
+    else:
+        out_shape = np.broadcast(
+            *[np.empty(a.shape, dtype=[]) for a in arrs]
+        ).shape
+        out_dtype = getattr(tiny_res, "dtype", arrs[0].dtype)
+        # Tempfile-backed memmap keeps result on disk
+        tmp = tempfile.NamedTemporaryFile(
+            prefix=f"{name}_", suffix=".dat", delete=False
+        )
+        tmp_path = tmp.name
+        tmp.close()
+        try:
+            out = np.memmap(
+                tmp_path, mode="w+", dtype=out_dtype, shape=out_shape
+            )
[…]
