### 2025-09-12 5d74e80
reduced size test problems (for packaging)
