### 2025-09-12 2c08009
added grid variables in custom var GUI + protections
