### 2025-09-14 1488ebe
fixed references journals
