+++ Paper/paper.bib
@@ -33,7 +33,7 @@
 @ARTICLE{MelonFuksman2021,
        author = {{Melon Fuksman}, D. and {Klahr}, H. and {Flock}, M. and {Mignone}, A.},
         title = "{A Two-moment Radiation Hydrodynamics Scheme Applicable to Simulations of Planet Formation in Circumstellar Disks}",
-      journal = {\apj},
+      journal = {ApJ},
      keywords = {Radiative transfer, Protoplanetary disks, Planet formation, 1335, 1300, 1241, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2021,
         month = jan,
@@ -52,7 +52,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Muley2024a,
        author = {{Muley}, D. and {Melon Fuksman}, D. and {Klahr}, H.},
         title = "{Three-temperature radiation hydrodynamics with PLUTO: Thermal and kinematic signatures of accreting protoplanets}",
-      journal = {\aap},
+      journal = {A\&A},
      keywords = {methods: numerical, protoplanetary disks, planet-disk interactions, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2024,
         month = jul,
@@ -88,7 +88,7 @@ archivePrefix = {arXiv},
 @ARTICLE{PLUTO_2007,
        author = {{Mignone}, A. and {Bodo}, G. and {Massaglia}, S. and {Matsakos}, T. and {Tesileanu}, O. and {Zanni}, C. and {Ferrari}, A.},
         title = "{PLUTO: A Numerical Code for Computational Astrophysics}",
-      journal = {\apjs},
+      journal = {ApJS},
      keywords = {Hydrodynamics, Methods: Numerical, Magnetohydrodynamics: MHD, Relativity, Shock Waves, Astrophysics},
          year = 2007,
         month = may,
@@ -106,7 +106,7 @@ archivePrefix = {arXiv},
 @ARTICLE{PLUTO_2012,
        author = {{Mignone}, A. and {Zanni}, C. and {Tzeferacos}, P. and {van Straalen}, B. and {Colella}, P. and {Bodo}, G.},
         title = "{The PLUTO Code for Adaptive Mesh Computations in Astrophysical Fluid Dynamics}",
-      journal = {\apjs},
+      journal = {ApJS},
      keywords = {hydrodynamics, magnetohydrodynamics: MHD, methods: numerical, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Cosmology and Nongalactic Astrophysics, Physics - Computational Physics},
          year = 2012,
         month = jan,
@@ -125,7 +125,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mignone_etal_2012,
        author = {{Mignone}, A. and {Flock}, M. and {Stute}, M. and {Kolb}, S.~M. and {Muscianisi}, G.},
         title = "{A conservative orbital advection scheme for simulations of magnetized shear flows with the PLUTO code}",
-      journal = {\aap},
+      journal = {A\&A},
      keywords = {methods: numerical, accretion, accretion disks, protoplanetary disks, magnetohydrodynamics (MHD), turbulence, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2012,
         month = sep,
@@ -143,7 +143,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mignone2018,
        author = {{Mignone}, A. and {Bodo}, G. and {Vaidya}, B. and {Mattia}, G.},
[…]
