### 2025-09-14 f1b5eac
updates bugli and melon paper
