+++ Paper/paper.bib
@@ -325,17 +325,18 @@ archivePrefix = {arXiv},
 @ARTICLE{Buglietal2024,
        author = {{Bugli}, M. and {Lopresti}, E.~F. and {Figueiredo}, E. and {Mignone}, A. and {Cerutti}, B. and {Mattia}, G. and {Del Zanna}, L. and {Bodo}, G. and {Berta}, V.},
         title = "{Relativistic reconnection with effective resistivity: I. Dynamics and reconnection rate}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Physics - Plasma Physics},
-         year = 2024,
-        month = oct,
-          eid = {arXiv:2410.20924},
-        pages = {arXiv:2410.20924},
-          doi = {10.48550/arXiv.2410.20924},
+      journal = {A\&A},
+     keywords = {acceleration of particles, magnetic reconnection, magnetohydrodynamics (MHD), plasmas, relativistic processes, methods: numerical, Astrophysics - High Energy Astrophysical Phenomena, Physics - Plasma Physics},
+         year = 2025,
+        month = jan,
+       volume = {693},
+          eid = {A233},
+        pages = {A233},
+          doi = {10.1051/0004-6361/202452277},
 archivePrefix = {arXiv},
        eprint = {2410.20924},
  primaryClass = {astro-ph.HE},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv241020924B},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025A&A...693A.233B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
@@ -480,20 +481,19 @@ archivePrefix = {arXiv},
 
 @ARTICLE{MelonFuksman2025,
        author = {{Melon Fuksman}, D. and {Flock}, M. and {Klahr}, H. and {Mattia}, G. and {Muley}, D.},
-        title = "{Multidimensional half-moment multigroup radiative transfer. Improving moment-based thermal models of circumstellar disks}",
-      journal = {arXiv e-prints},
-     keywords = {Instrumentation and Methods for Astrophysics, Earth and Planetary Astrophysics, Astrophysics of Galaxies, Solar and Stellar Astrophysics},
+        title = "{Multidimensional half-moment multigroup radiative transfer: Improving moment-based thermal models of circumstellar disks}",
+      journal = {\aap},
+     keywords = {radiative transfer, methods: numerical, protoplanetary disks},
          year = 2025,
-        month = apr,
-          eid = {arXiv:2504.13999},
-        pages = {arXiv:2504.13999},
-          doi = {10.48550/arXiv.2504.13999},
-archivePrefix = {arXiv},
-       eprint = {2504.13999},
- primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2025arXiv250413999M},
+        month = sep,
+       volume = {701},
+          eid = {A97},
+        pages = {A97},
+          doi = {10.1051/0004-6361/202554994},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025A&A...701A..97M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @ARTICLE{Costa2025,
        author = {{Costa}, A. and {Bodo}, G. and {Tavecchio}, F. and {Rossi}, P. and {Coppi}, P. and {Sciaccaluga}, A. and {Boula}, S.},
