+++ Paper/paper.bib
@@ -33,7 +33,7 @@
 @ARTICLE{MelonFuksman2021,
        author = {{Melon Fuksman}, D. and {Klahr}, H. and {Flock}, M. and {Mignone}, A.},
         title = "{A Two-moment Radiation Hydrodynamics Scheme Applicable to Simulations of Planet Formation in Circumstellar Disks}",
-      journal = {ApJ},
+      journal = {The Astrophysical Journal},
      keywords = {Radiative transfer, Protoplanetary disks, Planet formation, 1335, 1300, 1241, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2021,
         month = jan,
@@ -143,7 +143,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mignone2018,
        author = {{Mignone}, A. and {Bodo}, G. and {Vaidya}, B. and {Mattia}, G.},
         title = "{A Particle Module for the PLUTO Code. I. An Implementation of the MHD-PIC Equations}",
-      journal = {ApJ},
+      journal = {The Astrophysical Journal},
      keywords = {acceleration of particles, instabilities, magnetohydrodynamics: MHD, methods: numerical, plasmas, shock waves, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2018,
         month = may,
@@ -255,7 +255,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Vaidya2018,
        author = {{Vaidya}, B. and {Mignone}, A. and {Bodo}, G. and {Rossi}, P. and {Massaglia}, S.},
         title = "{A Particle Module for the PLUTO Code. II. Hybrid Framework for Modeling Nonthermal Emission from Relativistic Magnetized Flows}",
-      journal = {ApJ},
+      journal = {The Astrophysical Journal},
      keywords = {acceleration of particles, methods: numerical, polarization, radiation mechanisms: non-thermal, relativistic processes, shock waves, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = oct,
@@ -354,7 +354,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Wangetal2024,
        author = {{Wang}, Jie-Shuang and {Reville}, Brian and {Rieger}, Frank M. and {Aharonian}, Felix A.},
         title = "{Acceleration of Ultra-high-energy Cosmic Rays in the Kiloparsec-scale Jets of Nearby Radio Galaxies}",
-      journal = {\apjl},
+      journal = {The Astrophysical Journal Letters,
      keywords = {Ultra-high-energy cosmic radiation, Radio galaxies, Relativistic jets, Magnetohydrodynamical simulations, High energy astrophysics, 1733, 1343, 1390, 1966, 739, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Astrophysics of Galaxies},
          year = 2024,
         month = dec,
@@ -445,7 +445,7 @@ archivePrefix = {arXiv},
 @ARTICLE{MattiaFendt2022,
        author = {{Mattia}, G. and {Fendt}, C.},
         title = "{Jets from Accretion Disk Dynamos: Consistent Quenching Modes for Dynamo and Resistivity}",
-      journal = {ApJ},
+      journal = {The Astrophysical Journal},
[…]
