+++ Paper/paper.bib
@@ -52,7 +52,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Muley2024a,
        author = {{Muley}, D. and {Melon Fuksman}, D. and {Klahr}, H.},
         title = "{Three-temperature radiation hydrodynamics with PLUTO: Thermal and kinematic signatures of accreting protoplanets}",
-      journal = {A\&A},
+      journal = {Astronomy and Astrophysics},
      keywords = {methods: numerical, protoplanetary disks, planet-disk interactions, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2024,
         month = jul,
@@ -70,7 +70,7 @@ archivePrefix = {arXiv},
 @ARTICLE{NUMPY2020,
        author = {{Harris}, Charles R. and {Millman}, K. Jarrod and {van der Walt}, St{\'e}fan J. and {Gommers}, Ralf and {Virtanen}, Pauli and {Cournapeau}, David and {Wieser}, Eric and {Taylor}, Julian and {Berg}, Sebastian and {Smith}, Nathaniel J. and {Kern}, Robert and {Picus}, Matti and {Hoyer}, Stephan and {van Kerkwijk}, Marten H. and {Brett}, Matthew and {Haldane}, Allan and {del R{\'\i}o}, Jaime Fern{\'a}ndez and {Wiebe}, Mark and {Peterson}, Pearu and {G{\'e}rard-Marchant}, Pierre and {Sheppard}, Kevin and {Reddy}, Tyler and {Weckesser}, Warren and {Abbasi}, Hameer and {Gohlke}, Christoph and {Oliphant}, Travis E.},
         title = "{Array programming with NumPy}",
-      journal = {\nat},
+      journal = {Nature},
      keywords = {Computer Science - Mathematical Software, Statistics - Computation},
          year = 2020,
         month = sep,
@@ -88,7 +88,7 @@ archivePrefix = {arXiv},
 @ARTICLE{PLUTO_2007,
        author = {{Mignone}, A. and {Bodo}, G. and {Massaglia}, S. and {Matsakos}, T. and {Tesileanu}, O. and {Zanni}, C. and {Ferrari}, A.},
         title = "{PLUTO: A Numerical Code for Computational Astrophysics}",
-      journal = {ApJS},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {Hydrodynamics, Methods: Numerical, Magnetohydrodynamics: MHD, Relativity, Shock Waves, Astrophysics},
          year = 2007,
         month = may,
@@ -106,7 +106,7 @@ archivePrefix = {arXiv},
 @ARTICLE{PLUTO_2012,
        author = {{Mignone}, A. and {Zanni}, C. and {Tzeferacos}, P. and {van Straalen}, B. and {Colella}, P. and {Bodo}, G.},
         title = "{The PLUTO Code for Adaptive Mesh Computations in Astrophysical Fluid Dynamics}",
-      journal = {ApJS},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {hydrodynamics, magnetohydrodynamics: MHD, methods: numerical, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Cosmology and Nongalactic Astrophysics, Physics - Computational Physics},
          year = 2012,
         month = jan,
@@ -125,7 +125,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Mignone_etal_2012,
        author = {{Mignone}, A. and {Flock}, M. and {Stute}, M. and {Kolb}, S.~M. and {Muscianisi}, G.},
[…]
