+++ Paper/paper.bib
@@ -354,7 +354,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Wangetal2024,
        author = {{Wang}, Jie-Shuang and {Reville}, Brian and {Rieger}, Frank M. and {Aharonian}, Felix A.},
         title = "{Acceleration of Ultra-high-energy Cosmic Rays in the Kiloparsec-scale Jets of Nearby Radio Galaxies}",
-      journal = {The Astrophysical Journal Letters,
+      journal = {The Astrophysical Journal Letters},
      keywords = {Ultra-high-energy cosmic radiation, Radio galaxies, Relativistic jets, Magnetohydrodynamical simulations, High energy astrophysics, 1733, 1343, 1390, 1966, 739, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Astrophysics of Galaxies},
          year = 2024,
         month = dec,
@@ -482,7 +482,7 @@ archivePrefix = {arXiv},
 @ARTICLE{MelonFuksman2025,
        author = {{Melon Fuksman}, D. and {Flock}, M. and {Klahr}, H. and {Mattia}, G. and {Muley}, D.},
         title = "{Multidimensional half-moment multigroup radiative transfer: Improving moment-based thermal models of circumstellar disks}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {radiative transfer, methods: numerical, protoplanetary disks},
          year = 2025,
         month = sep,
