### 2025-09-14 #8448 @GiMattia
Hi @warrickball, I'm sorry, but it's a bit unclear to me what should be done now. The tag v 1.0.0 was referring to the repository at the time of the submission. Thanks to the reviewers' comments, the code has changed, so I created a new GitHub release, which I linked to a Zenodo DOI. I understand now that the issue is that the release tag should match the one at the time of the submission. Is that correct? In this case, I think I can simply change the GitHub release tag. From the "Review Process" part of the documentation, it seems that the issue tag should change, and not the GitHub release.
