### 2025-09-14 #8448 @warrickball
Ah, yes. There are two things at odds here. First, as you rightly say, the editor here should update the metadata at the end of the review with the new version. I apologise that we missed this at first. I'll do that as soon as we decide on a version number.

Second, I didn't realise that you've actually called the tagged release `PROOFS`. The version number in the `pyproject.toml` file is v1.1, which is what would ideally be released. I've recently found the Python ecosystem is becoming strict about version numbers according to [the Python packaging specifications](https://packaging.python.org/en/latest/specifications/version-specifiers/) but the logic is reasonable: it's hard to determine which versions match requirements unless they have a reasonable numerical order.

So I think the best solution here is for you to make a tagged release for v1.1.0 (or v1.1, or anything after v1.0.0, as long as the `pyproject.toml` specification is appropriate) and I'll update the JOSS metadata to match.
