+++ build/lib/pyPLUTO/imagefuncs/streamplot.py
@@ -1,0 +1,294 @@
+import warnings
+from typing import Any
+
+import numpy as np
+from matplotlib.collections import LineCollection
+from numpy.typing import NDArray
+
+from ..imagemixin import ImageMixin
+from ..imagestate import ImageState
+from ..utils.inspector import track_kwargs
+from .colorbar import ColorbarManager
+from .imagetools import ImageToolsManager
+from .range import RangeManager
+from .set_axis import AxisManager
+
+
+class StreamplotManager(ImageMixin):
+    """Manages the streamplot function from matplotlib.pyplot."""
+
+    exposed_methods = ("streamplot",)
+
+    def __init__(self, state: ImageState):
+        """Initialize the StreamplotManager with the given state."""
+        self.state = state
+
+        self.AxisManager = AxisManager(state)
+        self.ColorbarManager = ColorbarManager(state)
+        self.ImageToolsManager = ImageToolsManager(state)
+        self.RangeManager = RangeManager(state)
+
+    @track_kwargs
+    def streamplot(
+        self,
+        var1: NDArray[np.generic],
+        var2: NDArray[np.generic],
+        check: bool = True,
+        **kwargs: Any,
+    ) -> LineCollection:
+        """Plots a streamplot of a vector field. The function uses the
+        streamplot function from matplotlib.pyplot.
+
+        Returns
+        -------
+        - strm: LineCollection
+            The streamplot of the given vector field.
+
+        Parameters
+        ----------
+        - alpha: float, default 1.0
+            Sets the opacity of the plot, where 1.0 means total opaque and 0.0
+            means total transparent.
+        - arrowsize: float, default 1.0
+            Sets the size of the arrows of the streamline.
+        - arrowstyle: str, default '-|>'
+            Sets the style of the arrows of the streamline.
+        - aspect: {'auto', 'equal', float}, default 'auto'
+            Sets the aspect ratio of the plot. The 'auto' keyword is the default
+            option (most likely the plot will be squared). The 'equal' keyword
+            will set the same scaling for x and y. A float will fix the ratio
+            between the y-scale and the x-scale (1.0 is the same as 'equal').
+        - ax: axis object
+            The axis where to plot the streamlines. If not given, the last
+            considered axis will be used.
+        - brokenlines: bool, default True
+            Splits the streamlines in shorter segments.
+        - c: str, default self.color
+            Determines the streamplot color. If not defined, the program will
+            loop over an array of 6 color which are different for the most
+            common vision deficiencies.
+        - cmap: str, default 'hot'
+            Selects the colormap. If not defined, the colormap 'hot' will be
+            adopted. Some useful colormaps are: plasma, magma, seismic. Please
+            avoid using colorbars like jjet or rainbow, which are not
+            perceptively uniform and not suited for people with vision
+            deficiencies.
+            All the colormap available are listed in the following link:
+            https://matplotlib.org/stable/tutorials/colors/colormaps.html
+        - cpos: {'top','bottom','left','right'}, default None
+            Enables the colorbar (if defined), default position on the right.
+        - cscale: {'linear','log','symlog','twoslope'}, default 'linear'
+            Sets the colorbar scale. Default is the linear ('norm') scale.
+        - density: float, default 1.0
+            Sets the density and closeness of the streamlines. The domain is
+            divided in a 30x30 grid. When set as default, each cell contains at
+            most a number of crossing streamplot line equal to this keyword.
+        - extend: {'neither','both','min','max'}, default 'neither'
+            Sets the extension of the triangular colorbar extension.
+        - extendrect: bool, default False
+            If True, the colorbar extension will be rectangular.
+        - fontsize: float, default 17.0
[…]
