+++ build/lib/pyPLUTO/imagefuncs/set_axis.py
@@ -1,0 +1,396 @@
+import warnings
+from collections.abc import Iterable
+from typing import Any
+
+import matplotlib.pyplot as plt
+from matplotlib.axes import Axes
+
+from ..imagemixin import ImageMixin
+from ..imagestate import ImageState
+from ..utils.inspector import track_kwargs
+from .imagetools import ImageToolsManager
+from .range import RangeManager
+
+
+class AxisManager(ImageMixin):
+    """This class manages the axis of the image. It allows to customize the
+    axis of the image, such as the range, scale and aspect."""
+
+    exposed_methods = ("set_axis",)
+
+    def __init__(self, state: ImageState) -> None:
+        """Initializes the AxisManager with the given state."""
+        self.state = state
+        self.ImageToolsManager = ImageToolsManager(state)
+        self.RangeManager = RangeManager(state)
+
+    @track_kwargs
+    def set_axis(
+        self, ax: Axes | None = None, check: bool = True, **kwargs: Any
+    ) -> None:
+        """Customization of a single subplot axis. Properties such as the
+        range, scale and aspect of each subplot should be customized here.
+
+        Returns
+        -------
+        - None
+
+        Parameters
+        ----------
+        - alpha: float, default 1.0
+            Sets the opacity of the plot, where 1.0 means total opaque and 0.0
+            means total transparent.
+        - aspect: {'auto', 'equal', float}, default 'auto'
+            Sets the aspect ratio of the plot. The 'auto' keyword is the default
+            option (most likely the plot will be squared). The 'equal' keyword
+            will set the same scaling for x and y. A float will fix the ratio
+            between the y-scale and the x-scale (1.0 is the same as 'equal').
+        - ax: ax object, default None
+            The axis to customize. If None the current axis will be selected.
+        - fontsize: float, default 17.0
+            Sets the fontsize for all the axis components (only for the current
+            axis).
+        - grid: bool | string, default False
+            Enables/disables the grid on the plot. If True it enables both axes
+            grids. If 'x' or 'y' it enables only the x- or y-axis grid.
+        - labelsize: float, default fontsize
+            Sets the labels fontsize (which is the same for both labels).
+            The default value corresponds to the value of the keyword
+            'fontsize'.
+        - minorticks: str, default None
+            If not None enables the minor ticks on the plot (for both grid
+            axes).
+        - sharex: Matplotlib axis | False, default False
+            Shares the x-axis with another axis.
+        - sharey: Matplotlib axis | False, default False
+            Shares the y-axis with another axis.
+        - ticksdir: {'in', 'out'}, default 'in'
+            Sets the ticks direction. The default option is 'in'.
+        - tickssize: float | bool, default True
+            Sets the ticks fontsize (which is the same for both grid axes).
+            The default value corresponds to the value of the keyword
+            'fontsize'.
+        - title: str, default None
+            Places the title of the plot on top of it.
+        - titlepad: float, default 8.0
+            Sets the distance between the title and the top of the plot
+        - titlesize: float, default fontsize
+            Sets the title fontsize. The default value corresponds to the value
+            of the keyword 'fontsize'.
+        - xrange: [float, float], default [0,1]
+            Sets the range in the x-direction. If not defined the code will
+            compute the range while plotting the data.
+        - xscale: {'linear','log'}, default 'linear'
+            If enabled (and different from True), sets automatically the scale
[…]
