+++ build/lib/pyPLUTO/gui/custom_var.py
@@ -1,0 +1,427 @@
+"""A custom variable evaluator for PyPLUTO."""
+
+import os
+import re
+import tempfile
+
+import numexpr as ne
+import numpy as np
+from PyQt6.QtCore import Qt
+from PyQt6.QtWidgets import (
+    QComboBox,
+    QDialog,
+    QDialogButtonBox,
+    QFormLayout,
+    QLabel,
+    QPlainTextEdit,
+    QVBoxLayout,
+)
+
+SENTINEL = "Custom var..."
+
+# --- tiny normalizer ---------------------------------------------------------
+# Allow both D.something and bare names; allow np./numpy. prefixes.
+_NORM_PATTERNS = [
+    (re.compile(r"\bD\."), ""),  # D.Bx1 -> Bx1
+    (re.compile(r"\bnp\."), ""),  # np.sqrt -> sqrt
+    (re.compile(r"\bnumpy\."), ""),  # numpy.sqrt -> sqrt
+]
+
+
+def _normalize_expr(expr: str) -> str:
+    s = expr.strip()
+    for pat, repl in _NORM_PATTERNS:
+        s = pat.sub(repl, s)
+    return s
+
+
+def _frozen_var_names(D):
+    """Names that must never be overridden by custom vars."""
+    names = set(map(str, getattr(D, "_load_vars", [])))  # loaded-from-file vars
+    names.update({"x1", "x2", "x3"})  # base grid
+    g = (getattr(D, "geom", "") or "").upper()
+    if g == "CARTESIAN":
+        names.update({"x", "y", "z"})
+    elif g == "POLAR":
+        names.update({"R", "phi", "z", "x", "y", "xr", "yr"})
+    elif g == "SPHERICAL":
+        names.update({"r", "theta", "phi", "R", "z", "rt", "zt"})
+    return names
+
+
+def _apply_grid_shaping(local, D):
+    """
+    Reshape only x1/x2/x3 if they are 1-D so they broadcast with D.nshp.
+    PyPLUTO order: 1D -> (nx1,), 2D -> (nx1, nx2), 3D -> (nx1, nx2, nx3)
+    x1 aligns to axis 0, x2 to axis 1, x3 to axis 2.
+    """
+    nshp = D.nshp if isinstance(D.nshp, tuple) else tuple(D.nshp)
+    for axis, name in enumerate(("x1", "x2", "x3")[: D.dim]):
+        v = local.get(name)
+        if (
+            isinstance(v, np.ndarray)
+            and v.ndim == 1
+            and v.shape[0] == nshp[axis]
+        ):
+            try:
+                # make a singleton shape on other axes, then broadcast to D.nshp
+                pattern = [1] * D.dim
+                pattern[axis] = v.shape[0]  # axis-aligned length
+                v_view = v.reshape(tuple(pattern))
+                local[name] = np.broadcast_to(
+                    v_view, nshp
+                )  # FINAL SHAPE == D.nshp
+            except Exception:
+                pass  # stay quiet per your policy
+
+    if D.geom == "CARTESIAN":
+        local["x"] = local["x1"]
+        local["y"] = local["x2"]
+        local["z"] = local["x3"]
+    elif D.geom == "POLAR":
+        local["R"] = local["x1"]
+        local["phi"] = local["x2"]
+        local["z"] = local["x3"]
+        local["x"] = D.x1c.T[:, :, None] if D.dim == 3 else D.x1c.T
+        local["y"] = D.x2c.T[:, :, None] if D.dim == 3 else D.x2c.T
+    elif D.geom == "SPHERICAL":
+        local["r"] = local["x1"]
+        local["theta"] = local["x2"]
+        local["phi"] = local["x3"]
[…]
