+++ build/lib/pyPLUTO/image.py
@@ -1,70 +1,54 @@
-# image_new.py
+"""Image class. It plots the data."""
+
 from typing import Any
 
-from matplotlib.figure import Figure
-
 from .amr import oplotbox
-from .colorbar import ColorbarManager
-from .contour import ContourManager
-from .create_axes import CreateAxesManager
-from .delegator import delegator
-from .display import DisplayManager
-from .figure import FigureManager
+from .imagefuncs.colorbar import ColorbarManager
+from .imagefuncs.contour import ContourManager
+from .imagefuncs.create_axes import CreateAxesManager
+from .imagefuncs.display import DisplayManager
+from .imagefuncs.figure import FigureManager
+from .imagefuncs.imagetools import ImageToolsManager
+from .imagefuncs.interactive import InteractiveManager
+from .imagefuncs.legend import LegendManager
+from .imagefuncs.plot import PlotManager
+from .imagefuncs.range import RangeManager
+from .imagefuncs.scatter import ScatterManager
+from .imagefuncs.set_axis import AxisManager
+from .imagefuncs.streamplot import StreamplotManager
+from .imagefuncs.zoom import ZoomManager
 from .imagemixin import ImageMixin
 from .imagestate import ImageState
-from .imagetools import ImageToolsManager
-from .inspector import track_kwargs
-from .interactive import InteractiveManager
-from .legend import LegendManager
-from .mediator import Mediator
-from .plot import PlotManager
-from .scatter import ScatterManager
-from .set_axis import AxisManager
-from .streamplot import StreamplotManager
-from .zoom import ZoomManager
-
-manager_classes = (
-    AxisManager,
-    ColorbarManager,
-    ContourManager,
-    CreateAxesManager,
-    DisplayManager,
-    ImageToolsManager,
-    InteractiveManager,
-    LegendManager,
-    PlotManager,
-    ScatterManager,
-    StreamplotManager,
-    ZoomManager,
-)
+from .utils.inspector import track_kwargs
 
 
-@delegator("state", "mediator")
 class Image(ImageMixin):
-    """Image class. It plots the data. The Image class is a facade for the
-    different managers that handle the various aspects of plotting, such as
-    creating axes, displaying data, adding colorbars, and more. It provides a
-    unified interface for creating and managing plots in a figure.
-    The attributes are handled through the `ImageState` class, which is a
-    dataclass that stores the state of the image, such as the figure, axes,
-    and other properties. The `Image` class uses a mediator pattern to manage
-    the interactions between the different managers and the state."""
+    """Description of the Image class.
+
+    The Image class is a facade for the different managers that handle the
+    various aspects of plotting, such as creating axes, displaying data, adding
+    legends, text, fieldlines, colorbars, and more. It provides a unified
+    interface for creating and managing plots in a figure. The attributes are
+    handled through the `ImageState` class, which is a dataclass that stores the
+    state of the image, such as the figure, axes, and other properties. The
+    `Image` class uses a mediator pattern to manage the interactions between the
+    different managers and the state.
+
+    """
 
     @track_kwargs
     def __init__(
         self,
-        LaTeX: bool | str = True,
-        fig: Figure | None = None,
-        style: str = "default",
         text: bool = True,
         check: bool = True,
         **kwargs: Any,
     ) -> None:
-        """Initialization of the Image class that creates a new figure and sets
-        the LaTeX conditions, as well as the matplotlib style. Every Image is
-        associated to a figure object and only one in order to avoid confusion
-        between images and figures. If you want to create multiple figures, you
-        have to create multiple Image objects.
+        """Initialize the Image class.
+
+        Ihat creates a new figure and sets the LaTeX conditions, as well as the
+        matplotlib style. Every Image is associated to a figure object and only
[…]
