+++ build/lib/pyPLUTO/imagestate.py
@@ -9,21 +9,20 @@ from matplotlib.figure import Figure
 
 @dataclass
 class ImageState:
-    """Class that stores the state of the Image class. Its purpose is to
-    keep track of the current state of the image, such as the figure, axes,
-    and other properties and update the key attributes through all the
-    different classes that handle the plotting and display of the image at
-    runtime.
-    """
+    """Class that stores the state of the Image class.
 
-    LaTeX: bool | str
-    style: str
+    Its purpose is to keep track of the current state of the image, such as the
+    figure, axes, and other properties and update the key attributes through all
+    the different classes that handle the plotting and display of the image at
+    runtime."""
+
     ax: list[Axes] = field(default_factory=list)
     color: list[str] = field(default_factory=list)
     dictcol: dict[int, str] = field(default_factory=dict)
     fig: Figure | None = None
     figsize: list[float] = field(default_factory=lambda: [8.0, 5.0])
     fontsize: int = 17
+    LaTeX: bool | str = True
     legpar: list[list[float]] = field(default_factory=list)
     legpos: list[int | str | None] = field(default_factory=list)
     ncol0: int = 0
@@ -35,15 +34,16 @@ class ImageState:
     setay: list[Any | int] = field(default_factory=list)
     set_size: bool = False
     shade: list[str] = field(default_factory=list)
+    style: str = "default"
     tickspar: list[Any | int] = field(default_factory=list)
     tight: bool = True
     vlims: list[list[float]] = field(default_factory=list)
     xscale: list[str] = field(default_factory=list)
     yscale: list[str] = field(default_factory=list)
 
-    def __setattr__(self, name: str, value: object) -> None:
-        """Custom setter for the attributes of the ImageState class. It allows
-        to set the attributes of the class and update the state of the image.
-        This is useful for keeping track of the current state of the image and
-        updating it when necessary."""
-        super().__setattr__(name, value)
+    # def __setattr__(self, name: str, value: object) -> None:
+    #    """Custom setter for the attributes of the ImageState class. It allows
+    #    to set the attributes of the class and update the state of the image.
+    #    This is useful for keeping track of the current state of the image and
+    #    updating it when necessary."""
+    #    super().__setattr__(name, value)

[…]
