### 2025-09-15 #8448 @warrickball
Yes, I think the most complete thing to do is

* make a tagged GitHub release of version 1.1,
* create a Zenodo archive that corresponds to that release and
* post the Zenodo DOI here so I can update the metadata.

Sorry that I wasn't clear at first that the new tagged release (~v1.1) should be uploaded to Zenodo too. And don't worry at all about asking as many questions as you like!
