### 2025-09-16 35ca3c5
found weird repetition in paper
