+++ pyproject.toml
@@ -8,7 +8,7 @@ version = "1.1.1"
 description = "PyPLUTO: Plotting routines for PLUTO"
 readme = "README.md"
 requires-python = ">=3.10"
-license = "BSD-3-Clause"
+license = { text = "BSD-3-Clause" }
 authors = [
     { name = "G. Mattia", email = "mattia@mpia.de" },
     { name = "D. Crocco" },
