+++ build/lib/pyPLUTO/figure.py
@@ -1,391 +1,0 @@
-# figure_new.py
-import shutil
-import warnings
-from typing import Any
-
-import matplotlib as mpl
-import matplotlib.pyplot as plt
-from matplotlib.figure import Figure
-
-from .delegator import delegator
-from .imagestate import ImageState
-from .inspector import track_kwargs
-
-
-@delegator("state")
-class FigureManager:
-    """Manages the figure and sets the style, size, and LaTeX settings."""
-
-    @track_kwargs
-    def __init__(
-        self,
-        state: ImageState,
-        **kwargs: Any,
-    ) -> None:
-        """Initialization of the FigureManager class that creates a new
-        figure and sets the LaTeX conditions, as well as the matplotlib
-        style. Every Image is associated to a figure object and only one
-        in order to avoid confusion between images and figures. If you
-        want to create multiple figures, you have to create multiple
-        Image objects.
-
-        Returns
-        -------
-        - None
-
-        Parameters
-        ----------
-        - state: ImageState
-            The state of the image, which contains the figure and other
-            properties.
-        - kwargs: dict[str, Any]
-            Additional keyword arguments to customize the figure, such as
-            `figsize`, `fontsize`, `nwin`, `suptitle`, etc.
-
-        """
-        # needed because the __init__ is longer than simply self.state = state
-        self.state = state
-
-        # Extract specific kwargs for colorlines, with defaults if not provided
-        if "numcolor" in kwargs:
-            warnings.warn(
-                "numcolor is deprecated. Use numcolors instead.",
-                DeprecationWarning,
-            )
-
-        close = kwargs.pop("close", True)
-        fontweight = kwargs.pop("fontweight", "normal")
-        numcolors = kwargs.pop("numcolor", 10)  # DEPRECATED!!!
-        numcolors = kwargs.pop("numcolors", numcolors)
-        replace = kwargs.pop("replace", False)
-        suptitle = kwargs.pop("suptitle", None)
-        suptitlesize = kwargs.pop("suptitlesize", "large")
-        withblack = kwargs.pop("withblack", False)
-        withwhite = kwargs.pop("withwhite", False)
-        self.state.nwin = kwargs.get("nwin", self.state.nwin)
-
-        self.check_previous_fig(close)
-
-        self.state.fontsize = kwargs.get("fontsize", self.state.fontsize)
-        self.state.tight = kwargs.get("tight", self.state.tight)
-        self.state.figsize = kwargs.get("figsize", self.state.figsize)
-        if "figsize" in kwargs:
-            self.state.set_size = True
-
-        self.setup_style()
-        self.state.color = self.choose_colorlines(
-            numcolors, withblack, withwhite
-        )
-        self.assign_LaTeX(fontweight)
-        self.create_figure(replace, suptitle, suptitlesize)
-
-    def setup_style(self) -> None:
-        """Sets the matplotlib style."""
-        try:
-            plt.style.use(self.state.style)
-        except OSError:
-            warn = f"Warning: Style '{self.state.style}' not found. \
-                Switching to 'default'"
-            warnings.warn(warn, UserWarning)
-            self.state.style = "default"
-
-    def choose_colorlines(
-        self, numcolors: int, withblack: bool, withwhite: bool
-    ) -> list[str]:
-        """Chooses the colors for the lines. The colors are taken from a
-        list of colors that are suitable for all types of color vision
-        deficiencies.
-
-        Returns
-        -------
-        - colors: list[str]
-            The list of colors for the lines.
-
-        Parameters
-        ----------
-        - numcolors: int, default 10
-            The number of colors.
-        - withblack: bool, default False
-            If True, the black color is used as first color.
-        - withwhite: bool default False
-            If True, the white color is used as first color.
-
-        ----
-
-        Examples
-        --------
[…]
