+++ build/lib/pyPLUTO/contour.py
@@ -1,250 +1,0 @@
-import warnings
-from typing import Any
-
-import numpy as np
-from matplotlib.contour import QuadContourSet
-from numpy.typing import NDArray
-
-from .colorbar import ColorbarManager
-from .delegator import delegator
-from .imagestate import ImageState
-from .imagetools import ImageToolsManager
-from .inspector import track_kwargs
-from .range import RangeManager
-from .set_axis import AxisManager
-
-
-@delegator("state")
-class ContourManager:
-    """Class to manage contour plots in the image.
-
-    This class provides methods to create contour plots of variables in the
-    image class. It allows for customization of the contour lines, colorbars,
-    and other properties."""
-
-    exposed_methods = ("contour",)
-
-    def __init__(self, state: ImageState):
-        """Initialize the ContourManager with the given state."""
-        self.state = state
-        self.AxisManager = AxisManager(state)
-        self.ColorbarManager = ColorbarManager(state)
-        self.ImageToolsManager = ImageToolsManager(state)
-        self.RangeManager = RangeManager(state)
-
-    @track_kwargs
-    def contour(
-        self, var: NDArray[np.generic], check: bool = True, **kwargs: Any
-    ) -> QuadContourSet:
-        """Plots a contour plot of a given variable. The function uses the
-        matplotlib.pyplot.contour function. The function returns None.
-
-        Returns
-        -------
-        - cnt: LineCollection
-            The set of contour lines of the given variable.
-
-        Parameters
-        ----------
-        - alpha: float, default 1.0
-            Sets the transparency of the contour lines.
-        - aspect: {'auto', 'equal', float}, default 'auto'
-            Sets the aspect ratio of the plot. The 'auto' keyword is the default
-            option (most likely the plot will be squared). The 'equal' keyword
-            will set the same scaling for x and y. A float will fix the ratio
-            between the y-scale and the x-scale (1.0 is the same as 'equal').
-        - ax: {ax object, 'old', None}, default None
-            The axis where to plot the lines. If None, a new axis is created.
-            If 'old', the last considered axis will be used.
-        - c: str, default self.color
-            Determines the contour lines plot. If not defined, the program will
-            loop over an array of 6 color which are different for the most
-            common vision deficiencies.
-        - cmap: str, default 'hot'
-            Selects the colormap. If not defined, the colormap 'hot' will be
-            adopted. Some useful colormaps are: plasma, magma, seismic. Please
-            avoid using colorbars like jjet or rainbow, which are not
-            perceptively uniform and not suited for people with vision
-            deficiencies.
-        - cpos: {'top','bottom','left','right'}, default None
-            Enables the colorbar (if defined), default position on the right.
-        - cscale: {'linear','log','symlog','twoslope'}, default 'linear'
-            Sets the colorbar scale. Default is the linear ('norm') scale.
-        - extend: {'neither','both','min','max'}, default 'neither'
-            Sets the extension of the triangular colorbar extension.
-        - extendrect: bool, default False
-            If True, the colorbar extension will be rectangular.
-        - fontsize: float, default 17.0
-            Sets the fontsize for all the axis components (only for the current
-            axis).
-        - grid: bool, default False
-            Enables/disables the grid on the plot.
-        - labelsize: float, default fontsize
-            Sets the labels fontsize (which is the same for both labels).
-            The default value corresponds to the value of the keyword
-            'fontsize'.
-        - levels: np.ndarray
-            The levels of the contour lines.
-        - minorticks: str, default None
-            If not None enables the minor ticks on the plot (for both grid
-            axes).
[…]
