### 2025-09-16 #8448 @warrickball
Thanks for working through that with me @GiMattia. I think the metadata now looks good. 👍 I have some other checks to work through and will probably propose some editorial tweaks in a PR on the PyPLUTO repository, hopefully today or tomorrow.
