### 2025-09-16 f93775a
Initial editorial tweaks for JOSS paper
