+++ Paper/paper.bib
@@ -32,7 +32,7 @@
 
 @ARTICLE{MelonFuksman2021,
        author = {{Melon Fuksman}, D. and {Klahr}, H. and {Flock}, M. and {Mignone}, A.},
-        title = "{A Two-moment Radiation Hydrodynamics Scheme Applicable to Simulations of Planet Formation in Circumstellar Disks}",
+        title = {A Two-moment Radiation Hydrodynamics Scheme Applicable to Simulations of Planet Formation in Circumstellar Disks},
       journal = {The Astrophysical Journal},
      keywords = {Radiative transfer, Protoplanetary disks, Planet formation, 1335, 1300, 1241, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2021,
@@ -87,7 +87,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{PLUTO_2007,
        author = {{Mignone}, A. and {Bodo}, G. and {Massaglia}, S. and {Matsakos}, T. and {Tesileanu}, O. and {Zanni}, C. and {Ferrari}, A.},
-        title = "{PLUTO: A Numerical Code for Computational Astrophysics}",
+        title = {PLUTO: A Numerical Code for Computational Astrophysics},
       journal = {The Astrophysical Journal Supplement Series},
      keywords = {Hydrodynamics, Methods: Numerical, Magnetohydrodynamics: MHD, Relativity, Shock Waves, Astrophysics},
          year = 2007,
@@ -105,7 +105,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{PLUTO_2012,
        author = {{Mignone}, A. and {Zanni}, C. and {Tzeferacos}, P. and {van Straalen}, B. and {Colella}, P. and {Bodo}, G.},
-        title = "{The PLUTO Code for Adaptive Mesh Computations in Astrophysical Fluid Dynamics}",
+        title = {The PLUTO Code for Adaptive Mesh Computations in Astrophysical Fluid Dynamics},
       journal = {The Astrophysical Journal Supplement Series},
      keywords = {hydrodynamics, magnetohydrodynamics: MHD, methods: numerical, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Cosmology and Nongalactic Astrophysics, Physics - Computational Physics},
          year = 2012,
@@ -142,7 +142,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Mignone2018,
        author = {{Mignone}, A. and {Bodo}, G. and {Vaidya}, B. and {Mattia}, G.},
-        title = "{A Particle Module for the PLUTO Code. I. An Implementation of the MHD-PIC Equations}",
+        title = {A Particle Module for the PLUTO Code. I. An Implementation of the MHD-PIC Equations},
       journal = {The Astrophysical Journal},
      keywords = {acceleration of particles, instabilities, magnetohydrodynamics: MHD, methods: numerical, plasmas, shock waves, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2018,
@@ -160,7 +160,7 @@ archivePrefix = {arXiv},
 }
 
 @software{PANDAS2020,
-    author       = {The pandas development team},
+    author       = {{The pandas development team}},
     title        = {pandas-dev/pandas: Pandas},
     month        = feb,
     year         = 2020,
@@ -205,7 +205,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{SCIPY_2020,
[…]
