### 2025-09-17 #8448 @GiMattia
@warrickball I merged the changes.
