### 2025-09-17 f80f904
RFC/DEP: use inifix internally for parsing ini files
