+++ Docs/Badges/pylint.txt
@@ -53,7 +53,7 @@ Src/pyPLUTO/load.py:132:4: R0912: Too many branches (22/12) (too-many-branches)
 Src/pyPLUTO/load.py:132:4: R0915: Too many statements (124/50) (too-many-statements)
 Src/pyPLUTO/load.py:387:12: W0707: Consider explicitly re-raising using 'except AttributeError as exc' and 'raise AttributeError(f"'Load' object has no attribute '{name}'") from exc' (raise-missing-from)
 Src/pyPLUTO/load.py:389:4: C0415: Import outside toplevel (amr._DataScanHDF5, amr._inspect_hdf5) (import-outside-toplevel)
-Src/pyPLUTO/load.py:390:4: C0415: Import outside toplevel (defpluto._read_defh, defpluto._read_plini) (import-outside-toplevel)
+Src/pyPLUTO/load.py:390:4: C0415: Import outside toplevel (defpluto._read_defh) (import-outside-toplevel)
 Src/pyPLUTO/load.py:391:4: C0415: Import outside toplevel (echo_load.echo_load) (import-outside-toplevel)
 Src/pyPLUTO/load.py:392:4: C0415: Import outside toplevel (findlines._check_var, findlines.find_contour, findlines.find_fieldlines) (import-outside-toplevel)
 Src/pyPLUTO/load.py:393:4: C0415: Import outside toplevel (fourier.fourier) (import-outside-toplevel)
