+++ pyproject.toml
@@ -26,7 +26,9 @@ dependencies = [
     "numexpr",
     "pandas",
     "h5py",
-    "PyQt6"
+    "PyQt6",
+    "inifix>=5.1.0",
+    "typing-extensions>=4.5.0; python_version<'3.13'",
 ]
 
 [project.optional-dependencies]
