### 2025-09-17 9b8ef9e
try pathdir load files
