+++ pyPLUTO/configure.py
@@ -40,7 +40,7 @@ class Configure:
         self.session: str = self._find_session()
         self._setup_handlers(colorwarn, colorerr)
         if greet and not Configure.greeted:
-            print(f"PyPLUTO version: {self.version}   session: {self.session}")
+            print(f"PyPLUTO version: {self.version}  session: {self.session}")
             Configure.greeted = True
 
     def _find_session(self) -> str:
