### 2025-09-17 #8448 @warrickball
Thanks! One last change I'd suggest, since we still have a moment, is to add a publisher to the Paraview reference. The [DOI](https://doi.org/10.1016/B978-012387582-2/50038-1) resolves [here](https://www.sciencedirect.com/science/article/pii/B9780123875822500381?via%3Dihub), where the BibTeX produced by the "Cite" link has "Butterworth-Heineman" as publisher.

Could you please change [line 185 of `paper.bib`](https://github.com/GiMattia/PyPLUTO/blob/master/Paper/paper.bib#L185)?
```diff
--- publisher = "Unknown Publisher",
+++ publisher = {Butterworth-Heinemann},
```

Don't worry about re-reading the fresh proof I'll generate; we only need to check that one field. (And sorry for not picking this up in my PR.)
