+++ Paper/paper.bib
@@ -182,7 +182,7 @@ language = "English",
 isbn = "9780123875822",
 pages = "717--731",
 booktitle = "Visualization Handbook",
-publisher = "Unknown Publisher",
+publisher = {Butterworth-Heinemann},
 }
 
 @ARTICLE{Puzzoni2021,
