### 2025-09-17 #8448 @GiMattia
Thank you very much for everything! I assume the paper will soon be available on the webpage as well (right now, it says 404, but I can download it).
Thank you again for the smooth review process and for your patience.
